### 2022-01-05 #4035 @whedon
[PRE REVIEW]: svFSI: A Multiphysics Package for Integrated Cardiac Modeling
**Submitting author:** @CZHU20 (<a href="http://orcid.org/0000-0002-1099-8893">Chi Zhu</a>)
**Repository:** <a href="https://github.com/SimVascular/svFSI" target ="_blank">https://github.com/SimVascular/svFSI</a>
**Version:** v1.0.0
