### 2022-01-31 #4035 @Kevin-Mattheus-Moerman
@chennachaos @JaroslavHron @albansouche @meg-simula @soudah @loumalouomega @philbucher @sbryngelson @RiccardoRossi would you be interested in reviewing this work _"svFSI: A Multiphysics Package for Integrated Cardiac Modeling"_ for [JOSS, the journal of open source software](https://joss.theoj.org)?

JOSS reviews and publishes open source software. The [review process](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) takes place on GitHub and focuses on a very short paper (see here: [:page_with_curl:](https://github.com/openjournals/joss-papers/blob/joss.04035/joss.04035/10.21105.joss.04035.pdf)) and the software itself (see here: [:floppy_disk:](https://github.com/SimVascular/svFSI) ).

This here is a pre-review issue where we assign reviewers. Once we have sufficient reviewers we will open a dedicated review issue. Here is an example of such an issue for a previous submission: https://github.com/openjournals/joss-reviews/issues/2089

Let me know if you have questions. Thanks!
