### 2022-01-31 ac5ee22
Bugs fixed in shell mechanics formulation
A few bugs were fixed for the nonlinear shell mechanics formulation
with 3-noded constant strain triangles. Tested on plate bending and
aortic-valve mechanics simulation.

Commit also includes few cosmetic changes.

TODO:
- Postprocessing element data for NURBS need to be updated
- Shells for higher order elements (tri6, quad8/quad9) need to be
verified
- Shell formulation should be upgraded for hyperelastic material models
