+++ Code/Source/svFSI/SHELLS.f
@@ -32,7 +32,7 @@
 !-----------------------------------------------------------------------
 !
 !     This routines is for solving nonlinear shell mechanics problem
-!     using linear triangle finite elements and IGA.
+!     using finite elements and IGA.
 !
 !--------------------------------------------------------------------
 
@@ -72,7 +72,6 @@
          dl  = 0._RKIND
          bfl = 0._RKIND
          DO a=1, eNoN
-            Ac = lM%IEN(a,e)
             IF (a .LE. lM%eNoN) THEN
                Ac = lM%IEN(a,e)
                ptr(a) = Ac
@@ -90,20 +89,19 @@
          END DO
 
          IF (lM%eType .EQ. eType_TRI3) THEN
-!        Triangles - constant strain, no numerical integration
-            CALL SHELLTRI(lM, e, eNoN, al, yl, dl, xl, bfl, ptr)
+!           Constant strain triangles, no numerical integration
+            CALL SHELLCST(lM, e, eNoN, al, yl, dl, xl, bfl, ptr)
 
-         ELSE IF (lM%eType .EQ. eType_NRB) THEN
-!        NURBS
+         ELSE
             lR = 0._RKIND
             lK = 0._RKIND
 
-!           Update shape functions
-            CALL NRBNNX(lM, e)
+!           Update shape functions for NURBS elements
+            IF (lM%eType .EQ. eType_NRB) CALL NRBNNX(lM, e)
 
 !           Gauss integration
             DO g=1, lM%nG
-               CALL SHELLNRB(lM, g, eNoN, al, yl, dl, xl, bfl, lR, lK)
+               CALL SHELL3D(lM, g, eNoN, al, yl, dl, xl, bfl, lR, lK)
             END DO
 
 !           Assembly
@@ -124,7 +122,8 @@
       RETURN
       END SUBROUTINE CONSTRUCT_SHELL
 !####################################################################
-      SUBROUTINE SHELLTRI (lM, e, eNoN, al, yl, dl, xl, bfl, ptr)
+!     Construct shell mechanics for constant strain triangle elements
+      SUBROUTINE SHELLCST (lM, e, eNoN, al, yl, dl, xl, bfl, ptr)
       USE COMMOD
       USE ALLFUN
       IMPLICIT NONE
@@ -133,14 +132,14 @@
       REAL(KIND=RKIND), INTENT(IN) :: al(tDof,eNoN), yl(tDof,eNoN),
      2   dl(tDof,eNoN), xl(3,eNoN), bfl(3,eNoN)
 
-      LOGICAL :: bFlag, setIt(3)
+      LOGICAL :: setIt(3)
       INTEGER(KIND=IKIND) :: i, j, k, a, b, g
       REAL(KIND=RKIND) :: rho, dmp, elM, nu, ht, T1, amd, afl, w, Jac0,
      2   Jac, ud(3), fb(3), nV0(3), nV(3), gCov0(3,2), gCnv0(3,2),
      3   gCov(3,2), gCnv(3,2), x0(3,eNoN), xc(3,eNoN), eLoc(3,3),
      4   Jm(2,2), Qm(3,3), Dm(3,3), Em(3), Eb(3), DEm(3), DEb(3),
[…]
