### 2022-02-03 #4035 @rubenzorrilla
Many thanks for the offer but I don't have time right now to do a proper review. In any case thanks to this I am now aware of JOSS.
