### 2022-02-11 #4118 @axel-loewe
Zhu et al. present svFSI, a simulation software for cardiac physiology. It covers multiple cardiac functions (physics): solid mechanics, hemodynamics and apparently cardiac electrophysiology. The effort of developing a software for this challenging coupled problem and making if publicly available is appreciated.

>  License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?
- While there is no `LICENSE` file, there is `Copyright-Simvasscular.txt` which holds a license. Renaming might help to make this information more easily findable. The text seems to be the MIT license, I recommend mentioning this explicitly and also adding this information to the metadate of the GitHub project.

> Contribution and authorship: Has the submitting author (@CZHU20) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
From the commit history, this does not seem to be the case (3 minor commits only). I am aware that the public commits may not be representative of the entire history of the project. Please comment on it.

> Installation: Does installation proceed as outlined in the documentation?
I did not manage to compile the code with clang 13.0 (Undefined identifiers in METISLib).
Do you see an option to increase compatibility with other compilers or to provide a container? In the current form, I cannot evaluate svFSI's functionality.

> Functionality: Have the functional claims of the software been confirmed?
svFSI is introduced as a multi-physics simulator including electrophysiology, tissue mechanics and blood dynamics. Unfortunately, I could not find any information on electrophysiology in the paper or the documentation (http://simvascular.github.io/docssvFSI.html). As such, I suggest to either document these features or to remove these aspects from the paper.
Later on I found this link at the bottom of the README file: http://simvascular.github.io/docsSimCardio.html#cep-modeling
It's not entirely clear to me how SimCardio relates to svFSI and I strongly recommend to provide a comprehensive overview of all relevant documentation for svFSI at a central place.

> Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
The website lists them but there is no automated way of installing them. Later on, I discovered more detailed information in the `INSTALL-DEPS.md` file in the repository. Would be good to sync this information with the web page documentation.
Given the high number of dependencies and very specific versions being referred to in the documentation, I encourage you to consider providing a containerized version.

> State of the field: Do the authors describe how this software compares to other commonly-used packages?
> Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
As coupling multiple physics is a complex problem and also multiplies the number of user-chosen parameters, types of boundary conditions, pre- and postprocessing steps etc., I assume that svFSI probably provides only a subset of features of established single-physics simulators. I believe it would be important for potential new users to learn about which features are supported.
The brief paper remains rather vagues and does not mention the concrete scope. The documentation I found on the website is a tutorial for setting up a fluid-structure simulation (which I believe will be valuable for new users). However, I could not find a comprehensive user manual. The `Documentation` folder only contains a Doxygen configuration file.
I feel that a proper user documentation is a prerequisite for svFSI to be really useful beyond the group of developers.

> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
Could not find such information. There are no CI pipelines implemented as GitHub actions. Cannot judge on external CI.

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
Could not find such information in the repository or the svFSI part of the SimVascular webpage.

> References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?
Line 27: Please change the reference to openCARP to this one: https://dx.doi.org/10.1016/j.cmpb.2021.106223

- Is the `Readme.md` file required? Seems it does not hold relevant information for users/developers and might be confused with `README.md`. Will also cause trouble on non-case-sensitive file systems.
