### 2022-03-11 #4118 @Kevin-Mattheus-Moerman
> @Kevin-Mattheus-Moerman Thanks for the reminder and sorry for the delay. I am afraid that it will take a little bit longer. This is my first time submitting to JOSS. May I ask if I can revise the manuscript based on first reviewer's comments now? Will this affect other reviewers? Thank you.

@CZHU20 yes you can keep responding/fixing as the comments/issues come in. There is no former review round system. So yes people do start addressing reviewer comments and continue to do so.
