+++ Code/Source/svFSI/SETBC.f
@@ -267,7 +267,7 @@
 !     Geting the contribution of Neu BC
       IF (BTEST(lBc%bType,bType_cpl) .OR.
      2    BTEST(lBc%bType,bType_RCR)) THEN
-         h(1) = lBc%g
+         h(1) = lBc%g ! g contains the updated pressures and flowrates from cplBC, genBC, or RCR
       ELSE
          IF (BTEST(lBc%bType,bType_gen)) THEN
 !     Using "hl" as a temporary variable here
@@ -297,12 +297,12 @@
       ELSE
          DO a=1, nNo
             Ac     = lFa%gN(a)
-            hg(Ac) = -h(1)*lBc%gx(a)
+            hg(Ac) = -h(1)*lBc%gx(a) ! what is gx(a)?
          END DO
       END IF
 
 !     Add Neumann BCs contribution to the LHS/RHS
-      IF (lBc%flwP) THEN
+      IF (lBc%flwP) THEN ! If follower pressure load
          CALL BNEUFOLWP(lFa, hg, Dg)
       ELSE
          CALL BASSEMNEUBC(lFa, hg, Yg)
@@ -921,9 +921,11 @@
       INTEGER(KIND=IKIND) iFa, ptr, iBc, iM
       REAL(KIND=RKIND) tmp
 
+!     If coupling scheme is implicit, calculate the coupling derivative (M matrix in Moghadam 2013)
+!     as well as the  new pressures and flowrates from cplBC/genBC
       IF (cplBC%schm .EQ. cplBC_I) THEN
          CALL CALCDERCPLBC
-      ELSE
+      ELSE !   Just compute the new pressure and flowrates from cplBC/genBC
          RCRflag = .FALSE.
          DO iBc=1, eq(iEq)%nBc
             iFa = eq(iEq)%bc(iBc)%iFa
@@ -933,11 +935,13 @@
                IF (.NOT.RCRflag) RCRflag = .TRUE.
             END IF
             IF (ptr .NE. 0) THEN
+!              Compute flow rates from 3D on Neumann boundaries
                IF (BTEST(eq(iEq)%bc(iBc)%bType,bType_Neu)) THEN
                   cplBC%fa(ptr)%Qo = Integ(msh(iM)%fa(iFa),Yo,1,nsd)
                   cplBC%fa(ptr)%Qn = Integ(msh(iM)%fa(iFa),Yn,1,nsd)
                   cplBC%fa(ptr)%Po = 0._RKIND
                   cplBC%fa(ptr)%Pn = 0._RKIND
+!              Compute avg pressures from 3D on Dirichlet boundaries
                ELSE IF (BTEST(eq(iEq)%bc(iBc)%bType,bType_Dir)) THEN
                   tmp = msh(iM)%fa(iFa)%area
                   cplBC%fa(ptr)%Po = Integ(msh(iM)%fa(iFa),Yo,nsd+1)/tmp
@@ -947,13 +951,14 @@
                END IF
             END IF
          END DO
+!        Call genBC or cplBC to get updated pressures and flowrates
          IF (cplBC%useGenBC) THEN
             CALL genBC_Integ_X('T')
          ELSE
             CALL cplBC_Integ_X(RCRflag)
          END IF
       END IF
-
+!     Set g variable to updated pressures and flowrates
       DO iBc=1, eq(iEq)%nBc
          iFa = eq(iEq)%bc(iBc)%iFa
          ptr = eq(iEq)%bc(iBc)%cplBCptr
@@ -963,7 +968,10 @@
       RETURN
       END SUBROUTINE SETBCCPL
 !--------------------------------------------------------------------
-!     cplBC derivative is calculated here
[…]
