+++ Code/Source/svFSI/FLUID.f
@@ -1310,6 +1310,9 @@
       hc  = h*nV + udn*u
 
 !     Here the loop is started for constructing left and right hand side
+!     Add Neumann BC contributions to residual (lR) and stiffness (lK).
+!     Only the backflow stabilization is explicitly added to the stiffness. The
+!     resistance is added in the linear solver procedure later
       IF (nsd .EQ. 2) THEN
          DO a=1, eNoN
             lR(1,a) = lR(1,a) - w*N(a)*hc(1)
@@ -1322,12 +1325,13 @@
          END DO
       ELSE
          DO a=1, eNoN
+!           Add contribution of backflow stabilization and Neumann/coupled BC
             lR(1,a) = lR(1,a) - w*N(a)*hc(1)
             lR(2,a) = lR(2,a) - w*N(a)*hc(2)
             lR(3,a) = lR(3,a) - w*N(a)*hc(3)
             DO b=1, eNoN
 !              ? Where is the coupling term added to the tangent matrix (Moghadam 2013 eq 27)?
-               T1 = wl*N(a)*N(b)*udn
+               T1 = wl*N(a)*N(b)*udn ! This is only the backflow stabilization contribution
                lK(1,a,b)  = lK(1,a,b)  - T1
                lK(6,a,b)  = lK(6,a,b)  - T1
                lK(11,a,b) = lK(11,a,b) - T1
