### 2022-05-12 ca1fa63
added comments for integrating over deformed config
