+++ Code/Source/svFSI/ALLFUN.f
@@ -246,9 +246,16 @@
 
          DO g=1, lFa%nG ! For each Gauss integration point
             IF (.NOT.isIB) THEN
+
+!              Set mvMsh to true so that normal and area calc in GNNB takes into account deformation
+               mvMsh = .TRUE.
+
 !              Returns a vector (n) at element e and Gauss point g on face lFa
 !              that is the normal weighted by Jac
                CALL GNNB(lFa, e, g, nsd-1, lFa%eNoN, lFa%Nx(:,:,g), n)
+
+               mvMsh = .FALSE.
+
             ELSE
                CALL GNNIB(lFa, e, g, n)
             END IF
@@ -259,9 +266,6 @@
                Ac = lFa%IEN(a,e)
                DO i=1, nsd
                   sHat = sHat + lFa%N(a,g)*s(i,Ac)*n(i)
-                  !if eq%struct OR eq%FSI OR OR eq%lelas
-!                 sHat = sHat + lFa*N(a,g)*s(i,Ac)*nFi(i)*Jac, where nFi(i) and J are defined in BNEUFOLWP()
-!                 these quantities depend on the displacements Dg(:,Ac)
                END DO
             END DO
 !     Now integrating. Add product of Gauss weight and function value
