### 2022-05-12 13fc941
adding comments surrounding mvMsh in normal calc
