### 2022-05-13 cf468f9
update GNNB(). Not compiling, but close
