### 2022-05-13 7217e12
add comment about ADDBCMUL()
