+++ Code/Source/svFSI/NN.f
@@ -1923,12 +1923,8 @@ c        N(8) = lx*my*0.5_RKIND
       DO a=1, eNoN
          Ac = msh(iM)%IEN(a,Ec)
          lX(:,a) = x(:,Ac) ! get nodal coordinates from x (of reference configuration mesh)
-         !WRITE(*,eq(cEq)%s)
-         !WRITE(*,eq(cEq)%e)
-!        If struct or ustruct, add new displacement to nodal coordinates to get current config
-!        Should I deform the geometry by the old displacement Do or the new displacement Dn?
-         IF (eq(cEq)%phys .EQ. phys_struct .OR. eq(cEq)%phys .EQ. phys_ustruct) lX(:,a) = lX(:,a) + Dn(1:3,Ac) 
-!        IF (mvMsh) lX(:,a) = lX(:,a) + Do(nsd+2:2*nsd+1,Ac) ! Why this range of the Do vector? I believe these components are the fluid mesh displacements
+!        I believe Do(nsd+2:2*nsd+1) are the fluid mesh displacement in FSI
+         IF (mvMsh) lX(:,a) = lX(:,a) + Do(nsd+2:2*nsd+1,Ac)
       END DO
 
 !     Calculating surface deflation
@@ -2001,5 +1997,157 @@ c        N(8) = lx*my*0.5_RKIND
 
       RETURN
       END SUBROUTINE GNNB
+
+!--------------------------------------------------------------------
+!     This routine returns a vector at element "e" and Gauss point
+!     "g" of face "lFa" that is the normal weigthed by Jac, i.e.
+!     Jac = SQRT(NORM(n)). The normal is the surface normal in the current 
+!     configuration, and the Jacobian is the Jacobian of the mapping from parent
+!     surface element to current configuration surface element
+      SUBROUTINE GNNBT(lFa, e, g, insd, eNoNb, Nx, n)
+      USE COMMOD
+      USE ALLFUN
+      IMPLICIT NONE
+      INTEGER(KIND=IKIND), INTENT(IN) :: e, g, insd, eNoNb
+      REAL(KIND=RKIND), INTENT(IN) :: Nx(insd,eNoNb)
+      REAL(KIND=RKIND), INTENT(OUT) :: n(nsd)
+      TYPE(faceType), INTENT(IN) :: lFa
+
+      INTEGER(KIND=IKIND) a, Ac, i, iM, Ec, b, Bc, eNoN
+      REAL(KIND=RKIND) v(nsd)
+
+      LOGICAL, ALLOCATABLE :: setIt(:)
+      INTEGER(KIND=IKIND), ALLOCATABLE :: ptr(:)
+      REAL(KIND=RKIND), ALLOCATABLE :: lX(:,:), xXi(:,:)
+
+      iM   = lFa%iM
+      Ec   = lFa%gE(e)
+      eNoN = msh(iM)%eNoN
+
+      ALLOCATE(lX(nsd,eNoN), ptr(eNoN), setIt(eNoN))
+
+!     Creating a ptr list that contains pointer to the nodes of elements
+!     that are at the face at the beginning of the list and the rest at
+!     the end
+      setIt = .TRUE.
+      DO a=1, eNoNb
+         Ac = lFa%IEN(a,e)
+         DO b=1, eNoN
+            IF (setIt(b)) THEN
+               Bc = msh(iM)%IEN(b,Ec)
+               IF (Bc .EQ. Ac) EXIT
+            END IF
+         END DO
+         IF (b .GT. eNoN) THEN
+            WRITE(*,'(A)')
+            WRITE(*,'(A)') "=========================================="
+            WRITE(*,'(A)') " ERROR: could not find matching face nodes"
[…]
