### 2022-05-13 #4118 @CZHU20
@Kevin-Mattheus-Moerman Dear Kevin, thank you for the reminder and I am glad we all survived :)
We are steadily making progress in addressing the reviewer's comments. We think one of the most important issue is that the reviewer couldn't build the software and evaluate its functionality. We have created a Dockerfile to automate the build process and to increase its cross-platform compatibility. The Docker repository is located here:
https://github.com/CZHU20/Docker-svFSI
We have successfully tested it on macOS Big Sur, Ubuntu 18.04 and Window 10 with WSL 2, and is in the process of merging it into the main svFSI repository.

@axel-loewe Thank you very much for your comments and sorry for the delay. Would you please try to build svFSI with the above Dockerfile and review its functions? We are still addressing your other comments, especially those on documentations. We really appreciate your patience.
