### 2022-05-13 2c565b9
more comments tracking down cplBC tangent contribution
Adding more comments during my search for how cplBC adds to the tangent matrix. I believe the integrals in Moghadam et al. eq. 27 are computed in BAFINI.f in the variable sV. However, these are computed just once at initialization, and they are computed using the weighted normal vector in the reference configuration (using GNNB())
