+++ Code/Source/svFSI/EQASSEM.f
@@ -286,36 +286,7 @@
                   CALL BSTRUCT2D(eNoN, w, N, Nx, dl, hl, nV, lR, lK)
                END IF
             END IF
-
-!           Adding code more or less copied from BAFINI.f FSILSINI(). I think in BNEUFOLWP()
-!           we have to update the integrals involved in the resistance BC
-!           using the deformed geometry. This integral is contained in sV
-!           sV = int_Gammat (Na * n_i)
-            IF (BTEST(lBc%bType,bType_res)) THEN ! If resistance BC (or cpl BC)
-               sV = 0._RKIND
-               IF (lFa%eType .EQ. eType_NRB) CALL NRBNNXB(msh(iM),lFa,e) ! If NURBS
-!              Changed this to GNNBT() instead of GNNB() in FSILSINI()
-!              Tet weighted normal vector in current config
-               CALL GNNBT(lFa, e, g, nsd-1, lFa%eNoN, lFa%Nx(:,:,g),
-     2               n) 
-               DO a=1, lFa%eNoN ! Loop over nodes  in element
-                  Ac = lFa%IEN(a,e) ! Extract global nodal index
-                  sV(:,Ac) = sV(:,Ac) + lFa%N(a,g)*lFa%w(g)*n ! Integral of shape function times weighted normal
-               END DO
-               DO a=1, lFa%nNo
-                  Ac       = lFa%gN(a)
-                  sVl(:,a) = sV(:,Ac)
-               END DO
-               lsPtr     = lsPtr + 1
-               lBc%lsPtr = lsPtr
-   !           Fills lhs%face(i) variables, including val if sVl exists
-               CALL FSILS_BC_CREATE(lhs, lsPtr, lFa%nNo, nsd, 
-     2         BC_TYPE_Neu, gNodes, sVl)
-            ELSE
-               lBc%lsPtr = 0
-            END IF
-            
-         END DO
+         END DO ! Loop over Gauss point
 
 !        Now doing the assembly part
 #ifdef WITH_TRILINOS
@@ -337,8 +308,158 @@
          END IF
 #endif
          DEALLOCATE(ptr, hl, xl, dl, N, Nxi, Nx, lR, lK)
-      END DO
+      END DO ! Loop over elements
+
+!     Now update integrals involved in resistance BC contribution to
+!     stiffness matrix to reflect deformed geometry. Since we are using
+!     the deformed geometry to compute the contribution of the pressure
+!     load to the residual vector, we must also use the deformed geometry
+!     to compute the contribution of the resistance BC to the tangent
+!     matrix
+      CALL FSILSUPD(lBc, lFa, lBc%lsPtr)
+      
 
       RETURN
       END SUBROUTINE BNEUFOLWP
+
+! ----------------------------------------------------------------------
+!     Update the integral involved in the resistance BC input to the 
+!     linear solver to take into account the deformed geometry.
+!     This integral is sV = int_Gammat (Na * n_i) (See Moghadam 2013 eq. 27.)
+!     which is eventually used in ADDBCMUL() to add the resistance BC
+!     contribution to the matrix-vector product of the tangent matrix
+!     and an arbitrary vector.
+!     This code was more or less copied from BAFINI.f -> FSILSINI(). The
+!     major difference is that I call GNNBT() to get the weighted normal
[…]
