### 2022-05-21 173dbee
support for cap mesh
Made changes to NN.f so that a cap mesh, which has surface element that do not lie on volume elements, is still processed and an area weighted normal vector is still calculated. This seems to work fine with 1 proc (sequential), but there is some issue in DISTRIBUTE.f where the number of elements on the cap face is set to 0 after partitioning the face among processors (although this doesn't always happen)
