+++ Code/Source/svFSI/VTKXML.f
@@ -99,14 +99,24 @@
       CALL loadVTK(vtp, fName, iStat)
       IF (iStat .LT. 0) err = "VTP file read error (init)"
 
+      std = "check1"
+
+
       CALL getVTK_numPoints(vtp, lFa%nNo, iStat)
       IF (iStat .LT. 0) err = "VTP file read error (num points)"
+
+      std = "check2"
 
       CALL getVTK_numElems(vtp, lFa%nEl, iStat)
       IF (iStat .LT. 0) err = "VTP file read error (num cells)"
 
+      std = "check3"
+      PRINT*, "lFa%nEl", lFa%nEl
+
       CALL getVTK_nodesPerElem(vtp, lFa%eNoN, iStat)
       IF (iStat .LT. 0) err = "VTP file read error (nodes per cell)"
+
+      std = "check4"
 
       ALLOCATE(lFa%x(nsd,lFa%nNo), tmpX(maxNSD,lFa%nNo))
       CALL getVTK_pointCoords(vtp, tmpX, iStat)
@@ -114,9 +124,13 @@
       lFa%x(:,:) = tmpX(1:nsd,:)
       DEALLOCATE(tmpX)
 
+      std = "check5"
+
       ALLOCATE(lFa%IEN(lFa%eNoN,lFa%nEl))
       CALL getVTK_elemIEN(vtp, lFa%IEN, iStat)
       IF (iStat .LT. 0) err = "VTP file read error (ien)"
+
+      std = "check6"
 
       ALLOCATE(lFa%gN(lFa%nNo))
       CALL getVTK_pointData(vtp, "GlobalNodeID", lFa%gN, iStat)
@@ -133,6 +147,8 @@
          END DO
       END IF
 
+      std = "check7"
+
       ALLOCATE(lFa%gE(lFa%nEl))
       CALL getVTK_elemData(vtp, "GlobalElementID", lFa%gE, iStat)
       IF (iStat .LT. 0) THEN
@@ -144,6 +160,8 @@
          lFa%gebc(1,:) = lFa%gE(:)
          lFa%gebc(2:1+lFa%eNoN,:) = lFa%IEN(:,:)
       END IF
+
+      std = "check8"
 
       CALL flushVTK(vtp)
 
