+++ Code/Source/svFSI/NN.f
@@ -1852,8 +1852,8 @@ c        N(8) = lx*my*0.5_RKIND
 !--------------------------------------------------------------------
 !     This routine returns a vector at element "e" and Gauss point
 !     "g" of face "lFa" that is the normal weigthed by Jac, i.e.
-!     Jac = SQRT(NORM(n)). Jacobian of mapping from parent surface element to 
-!     ref config. surface element?
+!     Jac = SQRT(NORM(n)), the Jacobian of mapping from parent surface element to 
+!     ref configuration surface element.
       SUBROUTINE GNNB(lFa, e, g, insd, eNoNb, Nx, n)
       USE COMMOD
       USE ALLFUN
@@ -1874,6 +1874,15 @@ c        N(8) = lx*my*0.5_RKIND
       Ec   = lFa%gE(e)
       eNoN = msh(iM)%eNoN
 
+!     If Ec = 0, then this face element does not lie on a volume element, and
+!     we should just compute the area weighted normal vector anyway.
+      IF (Ec .EQ. 0) THEN
+!         WRITE(*,'(A)') "Face element not on volume element."
+!         WRITE(*,'(A)') "Calculate normal vector anyway."
+         CALL GNNBSURF(lFa, e, g, insd, eNoNb, Nx, n)
+         RETURN
+      END IF
+
       ALLOCATE(lX(nsd,eNoN), ptr(eNoN), setIt(eNoN))
 
 !     Creating a ptr list that contains pointer to the nodes of elements
@@ -1888,7 +1897,7 @@ c        N(8) = lx*my*0.5_RKIND
                IF (Bc .EQ. Ac) EXIT
             END IF
          END DO
-         IF (b .GT. eNoN) THEN
+         IF (b .GT. eNoN) THEN ! This occurs if a face element does not lie on a volume element
             WRITE(*,'(A)')
             WRITE(*,'(A)') "=========================================="
             WRITE(*,'(A)') " ERROR: could not find matching face nodes"
@@ -1977,7 +1986,7 @@ c        N(8) = lx*my*0.5_RKIND
          xXi = 0._RKIND
 !        AB 5/11/22: How does this calculation work?
          DO a=1, eNoNb
-            b = ptr(a)
+            b = ptr(a) ! get local node index of the boundary element
             DO i=1, insd
                xXi(:,i) = xXi(:,i) + Nx(i,a)*lX(:,b)
             END DO
@@ -1988,9 +1997,9 @@ c        N(8) = lx*my*0.5_RKIND
 
 !     Changing the sign if neccessary. a locates on the face and b
 !     outside of the face, in the parent element
-      a = ptr(1)
-      b = ptr(lFa%eNoN+1)
-      v = lX(:,a) - lX(:,b)
+      a = ptr(1) ! a is a node that lies on the boundary
+      b = ptr(lFa%eNoN+1) ! b is a node that lies in the interior
+      v = lX(:,a) - lX(:,b) ! v points outward
       IF (NORM(n,v) .LT. 0._RKIND) n = -n
 
       DEALLOCATE(setIt, ptr, lX)
@@ -2024,6 +2033,15 @@ c        N(8) = lx*my*0.5_RKIND
       iM   = lFa%iM
       Ec   = lFa%gE(e)
       eNoN = msh(iM)%eNoN
+
+!     If Ec = 0, then this face element does not lie on a volume element, and
+!     we should just compute the area weighted normal vector anyway.
[…]
