+++ Code/Source/svFSI/READMSH.f
@@ -459,6 +459,11 @@ c               END IF
          END IF
       END IF
 
+      DO iM=1, nMsh
+         DO iFa=1, msh(iM)%nFa
+            PRINT*, "msh(iM)%fa(iFa)%nEl", msh(iM)%fa(iFa)%nEl
+         END DO
+      END DO
       RETURN
       END SUBROUTINE READMSH
 !####################################################################
