### 2022-05-21 f592636
Adding GNNBSURFT()
Computes weighted surface normal for surface element in deformed configuration. Same as GNNBSURF(), but uses current configuration nodal positions, analogous to GNNBT().
