+++ Code/Source/svFSI/NN.f
@@ -2039,7 +2039,7 @@ c        N(8) = lx*my*0.5_RKIND
       IF (Ec .EQ. 0) THEN
          !WRITE(*,'(A)') "Face element not on volume element."
          !WRITE(*,'(A)') "Calculate normal vector anyway."
-         CALL GNNBSURF(lFa, e, g, insd, eNoNb, Nx, n)
+         CALL GNNBSURFT(lFa, e, g, insd, eNoNb, Nx, n)
          RETURN
       END IF
 
@@ -2094,7 +2094,7 @@ c        N(8) = lx*my*0.5_RKIND
          Ac = msh(iM)%IEN(a,Ec)
          lX(:,a) = x(:,Ac) ! get nodal coordinates from x (of reference configuration mesh)
 !        IF (mvMsh) lX(:,a) = lX(:,a) + Do(nsd+2:2*nsd+1,Ac) ! Why this range of the Do vector? I believe these components are the fluid mesh displacements
-!        Should I deform the geometry by the old displacement Do or the new displacement Dn?
+!        Deform the geometry by the new displacements Dn?
          lX(:,a) = lX(:,a) + Dn(:,Ac) 
       END DO
 
@@ -2174,7 +2174,9 @@ c        N(8) = lx*my*0.5_RKIND
 !     "g" of face "lFa" that is the normal weigthed by Jac, i.e.
 !     Jac = SQRT(NORM(n)), the Jacobian of mapping from parent surface element to 
 !     ref configuration surface element.
-!     TO DO: Compute area weighted normal vector for surface element
+!     This function is called for face elements that do not lie on a volume element
+!     For these elements, the direction of the normal vector is assumed from the
+!     nodal ordering.
       SUBROUTINE GNNBSURF(lFa, e, g, insd, eNoNb, Nx, n)
       USE COMMOD
       USE ALLFUN
@@ -2207,7 +2209,7 @@ c        N(8) = lx*my*0.5_RKIND
       END DO
 
 !     Calculating surface deflation
-      IF (msh(iM)%lShl) THEN ! If mesh is a shell
+      IF (msh(iM)%lShl) THEN ! If mesh is a shell. I think this is unnecessary in this function
 !        Since the face has only one parametric coordinate (edge), find
 !        its normal from cross product of mesh normal and interior edge
 
@@ -2267,5 +2269,107 @@ c        N(8) = lx*my*0.5_RKIND
 
       RETURN
       END SUBROUTINE GNNBSURF
+
+!--------------------------------------------------------------------
+!     This routine returns a vector at element "e" and Gauss point
+!     "g" of face "lFa" that is the normal weigthed by Jac, i.e.
+!     Jac = SQRT(NORM(n)), the Jacobian of mapping from parent surface element to 
+!     current configuration surface element.
+!     This function is called for face elements that do not lie on a volume element
+!     For these elements, the direction of the normal vector is assumed from the
+!     nodal ordering.
+!     Same as GNNBSURF(), except uses current configuration nodal positions.
+      SUBROUTINE GNNBSURFT(lFa, e, g, insd, eNoNb, Nx, n)
+      USE COMMOD
+      USE ALLFUN
+      IMPLICIT NONE
+      INTEGER(KIND=IKIND), INTENT(IN) :: e, g, insd, eNoNb
[…]
