+++ Code/Source/svFSI/DISTRIBUTE.f
@@ -82,7 +82,9 @@
       CALL SPLITJOBS(nMsh, cm%np(), wgt, wrk)
 
 !     First partitioning the meshes
-!     gmtl:  gtnNo --> tnNo
+!     gmtl:  gtnNo --> tnNo. What is gmtl? I think gmtl is a mapping from
+!     global nodal index to nodal index local to this proc. If this proc does
+!     not own an element with global nodal index Ac, then gmtl(Ac) = 0.
       tnNo = 0
       gmtl = 0
       IF (cm%seq()) THEN
@@ -95,6 +97,7 @@
       DO iM=1, nMsh
          dbg = "Partitioning mesh "//iM
          iWgt = REAL(wgt(iM,:)/SUM(wgt(iM,:)), KIND=RKIND4)
+!        what is gmtl??
          CALL PARTMSH(msh(iM), gmtl, cm%np(), iWgt)
       END DO
 
@@ -156,10 +159,8 @@
       DO iM=1, nMsh
          ALLOCATE(tMs(iM)%fa(msh(iM)%nFa))
          DO iFa=1, msh(iM)%nFa
-            PRINT*, "before PARTFACE: ", iM, iFa, msh(iM)%fa(iFa)%nEl
             CALL PARTFACE(msh(iM), msh(iM)%fa(iFa), tMs(iM)%fa(iFa),
      2         gmtl)
-            PRINT*, "after PARTFACE: ", iM, iFa, msh(iM)%fa(iFa)%nEl
          END DO
       END DO
 
@@ -1107,7 +1108,7 @@
       IMPLICIT NONE
       INTEGER(KIND=IKIND), INTENT(IN) :: nP
       REAL(KIND=RKIND4), INTENT(IN) :: wgt(nP)
-      INTEGER(KIND=IKIND), INTENT(INOUT) :: gmtl(gtnNo)
+      INTEGER(KIND=IKIND), INTENT(INOUT) :: gmtl(gtnNo) ! gtnNo = global total number of nodes
       TYPE(mshType), INTENT(INOUT) :: lM
 
       LOGICAL :: flag, fnFlag
@@ -1120,14 +1121,15 @@
      2   tempIEN(:,:), gtlPtr(:), sCount(:), disp(:)
       REAL(KIND=RKIND), ALLOCATABLE :: tmpR(:), tmpFn(:,:)
 
+!     If sequential, copy global parameters as local parameters and return
       IF (cm%seq()) THEN
-         lM%nEl = lM%gnEl
-         lM%nNo = lM%gnNo
+         lM%nEl = lM%gnEl ! number of elements in this mesh
+         lM%nNo = lM%gnNo ! number of nodes in this mesh
          ALLOCATE(lM%IEN(lM%eNoN,lM%nEl), lM%eDist(0:cm%np()))
-         lM%IEN      = lM%gIEN
+         lM%IEN      = lM%gIEN ! IEN array (maps element e (of mesh) and node a to global node A (of mesh))
          lM%eDist(0) = 0
-         lM%eDist(1) = lM%gnEl
-         ALLOCATE(lM%otnIEN(lM%nEl))
+         lM%eDist(1) = lM%gnEl ! element distribution array, lM%eDist(i) represents first element which belong to cm%id()=i
+         ALLOCATE(lM%otnIEN(lM%nEl))! old to new IEN array (maps ??)
          DO e=1, lM%nEl
             lM%otnIEN(e) = e
          END DO
@@ -1156,7 +1158,7 @@
       IF (lM%lShl) insd = nsd - 1
       IF (lM%lFib) insd = 1
 
-      eNoN = lM%eNoN
+      eNoN = lM%eNoN ! number of nodes in the element
[…]
