+++ Code/Source/svFSI/DISTRIBUTE.f
@@ -1327,7 +1327,7 @@ c            wrn = " ParMETIS failed to partition the mesh"
          ALLOCATE(tempIEN(eNoN,lM%gnEl), lM%otnIEN(lM%gnEl))
 !     Making the lM%IEN array in order, based on the cm%id() number in
 !     master. lM%otnIEN maps old IEN order to new IEN order.
-!     Why does the IEN order change?
+!     Why does the IEN order change? I think to facilitate scattering later.
          disp = 0
          DO e=1, lM%gnEl ! Loop over global elements in mesh
 !           gPart(e) = proc that owns e
@@ -1420,15 +1420,15 @@ c            wrn = " ParMETIS failed to partition the mesh"
       DEALLOCATE(tempIEN)
 
 !     Constructing the initial global to local pointer
-!     lM%IEN: eNoN,nEl --> gnNo
-!     gtlPtr: gnNo     --> nNo
-!     lM%IEN: eNoN,nEl --> nNo
+!     lM%IEN: eNoN,nEl --> gnNo. Maps node a of element e (local to proc and mesh) to global node A on mesh
+!     gtlPtr: gnNo     --> nNo. Maps global node A on mesh to local node a on proc and mesh
+!     lM%IEN: eNoN,nEl --> nNo. Reconstructing IEN to map node a of element e (local to proc and mesh) to local node a on proc and mesh
       ALLOCATE(gtlPtr(lM%gnNo)) ! global number of nodes on this mesh
       nNo    = 0
       gtlPtr = 0
       DO e=1, nEl ! Loop over num elements on this proc on this mesh
          DO a=1, eNoN ! Loop over nodes in element
-            Ac = lM%IEN(a,e) ! Get global node Ac on this mesh
+            Ac = lM%IEN(a,e) ! Get global node Ac on this mesh. This is the global node ID in Paraview
 !           If Ac has not be seen yet, add it to this proc. Increment the number
 !           of nodes on this proc, and update gtlPtr
             IF (gtlPtr(Ac) .EQ. 0) THEN
@@ -1447,7 +1447,7 @@ c            wrn = " ParMETIS failed to partition the mesh"
 !     part:  nNo  --> gtnNo. Maps local node index on this proc on this mesh to global
       ALLOCATE(part(nNo))
       DO Ac=1, lM%gnNo ! Loop over nodes on this mesh
-         a = gtlPtr(Ac) ! get local node index on this proc 
+         a = gtlPtr(Ac) ! get local node index on this proc and mesh
          ! If a is non-zero, then this proc owns global node Ac, so assign the
          ! corresponding global total node gN(Ac) to part(a).
          ! At this point, gN(Ac) maps global node index on this mesh to
@@ -1457,7 +1457,7 @@ c            wrn = " ParMETIS failed to partition the mesh"
 !     mapping and converting other parameters.
 !     I will use an upper bound for gPart as a container for ltg,
 !     since there can be repeated nodes. gPart is just a temp variable.
-!     gmtl:  gtnNo --> tnNo. Maps global node index across all meshes to local node index (across all meshes or parts of meshes belong to this proc)
+!     gmtl:  gtnNo --> tnNo. Maps global node index across all meshes to local node index (across all meshes or parts of meshes belong to this proc). gmtl is zero global node Ac does not belong to this proc
 !     gPart: tnNo  --> gtnNo. Maps local node index on this proc (across multiple parts of meshes) to global node index across all meshes. Inverse of gmtl
 !     ltg:   tnNo  --> gtnNo. How is this different from gPart, if at all? I think they contain the same info, gPart is just temporary as we assemble the info
[…]
