+++ Code/Source/svFSI/ALLFUN.f
@@ -103,12 +103,13 @@
       REAL(KIND=RKIND) IntegS, snode
 
       LOGICAL isIB, flag
-      INTEGER(KIND=IKIND) a, e, g, Ac, nNo, insd
+      INTEGER(KIND=IKIND) a, e, g, Ac, nNo, insd, p, ierr
       REAL(KIND=RKIND) sHat, Jac, n(nsd)
       TYPE(fsType) :: fs
+      REAL(KIND=RKIND), ALLOCATABLE :: sgather(:)
 
       IF (lFa%virtual) THEN
-         PRINT*, 'proc: ', cm%id(), 'Face name: ', lFa%name
+         !PRINT*, 'proc: ', cm%id(), 'Face name: ', lFa%name
       END IF
 
       flag = .FALSE.
@@ -128,9 +129,13 @@
          END IF
       END IF
 
+!     If face is virtual, allocate sgather vector. This will hold
+!     snode values from all procs after MPI_Gather operation on master
       IF (lFa%virtual) THEN
+         !PRINT*, 'cm%np()', cm%np()
+         ALLOCATE(sgather(cm%np()))
          DO a=1, nNo
-            PRINT*, 'proc: ', cm%id(), 'nNo: ', nNo, 's(a): ', s(a)
+            !PRINT*, 'proc: ', cm%id(), 'nNo: ', nNo, 's(a): ', s(a)
          END DO
       END IF
       isIB = .FALSE.
@@ -170,7 +175,7 @@
       CALL MPI_BARRIER(cm%com(), err)
 
       IntegS = 0._RKIND
-      PRINT*, "nEl inside IntegS(): ", lFa%nEl
+      !PRINT*, "nEl inside IntegS(): ", lFa%nEl
       DO e=1, lFa%nEl
 !     Updating the shape functions, if this is a NURB
          IF (lFa%eType .EQ. eType_NRB) THEN
@@ -184,8 +189,8 @@
             fs%Nx = lFa%Nx
          END IF
 
-         IF (lFa%virtual) THEN
-            PRINT*, "e inside IntegS(): ", e
+         IF (lFa%virtual .AND. cm%mas()) THEN
+            !PRINT*, 'proc: ', cm%id(),  "e inside IntegS(): ", e
          END IF
 
          DO g=1, fs%nG
@@ -197,8 +202,9 @@
             END IF
             Jac = SQRT(NORM(n))
 
-            IF (lFa%virtual) THEN
-               PRINT*, "g inside IntegS(): ", g
+            IF (lFa%virtual .AND. cm%mas()) THEN
+               !PRINT*, "g inside IntegS(): ", g
+               !PRINT*, "n inside IntegS(): ", n(1), n(2), n(3)
             END IF
 
 !     Calculating the function value
@@ -207,12 +213,43 @@
 !              If face is virtual, then master must get the function nodal value
 !              s(Ac) from another proc.
                IF (.NOT. lFa%virtual) THEN
-                  Ac   = lFa%IEN(a,e) ! Get local node number on proc Ac in [1,tnNo]
-                  snode = s(Ac)
+                  Ac   = lFa%IEN(a,e) ! Get local node number on proc. Ac in [1,tnNo]
+                  snode = s(Ac) ! Get nodal function value to use
                ELSE
-                  PRINT*, "Ac inside IntegS(): ", Ac
-                  snode = 0
-               END IF     
+                  Ac   = lFa%IEN(a,e) ! Get local node number on proc. Ac in [1,tnNo]
+!                 Ac = 0 if this proc does not own this node
+                  IF (Ac .EQ. 0) THEN ! This proc doesn't own this node
+                     snode = 0._RKIND ! Set snode to zero
+                  ELSE ! This proc does own this node
[…]
