+++ Code/Source/svFSI/NN.f
@@ -2193,22 +2193,66 @@ c        N(8) = lx*my*0.5_RKIND
       LOGICAL, ALLOCATABLE :: setIt(:)
       INTEGER(KIND=IKIND), ALLOCATABLE :: ptr(:)
       REAL(KIND=RKIND), ALLOCATABLE :: lX(:,:), xXi(:,:)
+      REAL(KIND=RKIND), ALLOCATABLE :: tmpX(:), tmpXg(:,:)
+      INTEGER(KIND=IKIND) ierr, p
 
       iM   = lFa%iM
       Ec   = lFa%gE(e)
       eNoN = msh(iM)%eNoN
 
       ALLOCATE(lX(nsd,eNoN), ptr(eNoN), setIt(eNoN))
+      ! Arrays to hold nodal positions for MPI gather operation
+      ALLOCATE(tmpX(nsd), tmpXg(nsd,cm%np()))
 
-!     Correcting the position vector if mesh is moving
+!     Communicating nodal positions and correcting the position vector if mesh is moving
       DO a=1, eNoNb ! Loop over nodes of boundary surface element
-!         Ac = msh(iM)%IEN(a,Ec)
-         Ac = lFa%IEN(a,e)
-         PRINT*, 'Ac inside GNNBSURF', Ac
-         lX(:,a) = x(:,Ac) ! get nodal coordinates from x (of reference configuration mesh)
+         Ac   = lFa%IEN(a,e) ! Get local node number on proc. Ac in [1,tnNo]
+!        Ac = 0 if this proc does not own this node
+         IF (Ac .EQ. 0) THEN ! This proc doesn't own this node
+            tmpX = 0._RKIND ! Set position of this node to zero as a flag
+         ELSE ! This proc does own this node
+            tmpX(:) = x(:,Ac) ! Get position of this node
+         END IF
+         IF (cm%mas()) THEN
+!            PRINT*, 'Before gather master tmpX: ', 
+!     2              tmpX(1), tmpX(2), tmpX(3)
+         END IF
+!        Communicate the true tmpX value among all procs. We use the 
+!        gather operation to get tmpX from all procs onto master
+         ! First, allocate space for tmpX values. The size of tmpXg
+         ! array is the number of procs being used.
+         tmpXg = 0._RKIND
+         !PRINT*, 'proc: ', cm%id(), 'snode: ', snode
+         CALL MPI_GATHER(tmpX, nsd, mpreal, tmpXg, nsd,
+     2            mpreal, master, cm%com(), ierr)
+!        If master, go through tmpXg and search for first non-zero 
+!         value.
+         IF (cm%mas()) THEN
+            DO p=1, cm%np()+1
+               IF (.NOT. ALL(tmpXg(:,p) .EQ. 0)) THEN
+                  tmpX = tmpXg(:,p)
+                  EXIT 
+               END IF
+            END DO
+         ELSE ! If not master, set tmpX to 0
+            tmpX = 0._RKIND
+         END IF 
+         !PRINT*, 'Ac inside GNNBSURF', Ac
+         lX(:,a) = tmpX(:) ! get nodal coordinates from x (of reference configuration mesh)
+         IF (cm%mas()) THEN
+!            PRINT*, 'After gather master lX: ', 
+!     2              lX(1,a), lX(2,a), lX(3,a)
+         END IF
 !        I believe Do(nsd+2:2*nsd+1) are the fluid mesh displacement in FSI
          IF (mvMsh) lX(:,a) = lX(:,a) + Do(nsd+2:2*nsd+1,Ac)
       END DO
+
+!     If slave, don't perform calculation of n. Wait for master to complete
[…]

[…]
