### 2022-05-28 127deed
virtual face bug fixes and tangent matrix notes
In SETBC.f, added a line so that if a face is virtual and there is a Neumann or traction BC defined on it, it will cycle. Otherwise, this leads to a crash.

Also, when testing an idealized LV coupled to a constant flow rate LPN using genBC, I noticed the convergence was suboptimal (less than linear, when it should be quadratic). I think this is because the tangent matrix is inconsistent with the residual vector when a virtual face is used to compute flow rate. In this case, the integrals in the coupled BC tangent matrix contribution (see Moghadam
[…]
