### 2022-05-30 bbdae6a
Add build instruction for Docker image.
