### 2022-05-31 5f6d75f
virtual face tangent mat and debugging, not working
I believe when a virtual face is used to compute flow rate, the virtual face should have some contribution to the tangent matrix, even if not pressure is applied to it. Because no pressure is applied to it, the tangent matrix contribution should involve an integral over the capped surface and an integral over the uncapped surface. In ADDBCMUL(), I believe this amounts to ignoring the second do loop when the face is virtual. I attempted to do this, but I think the integral (in ValM), as well as the global indices in glob() may be incorrect whe
[…]
