+++ Code/Source/svFSILS/FSILS_API.h
@@ -99,13 +99,15 @@
          END SUBROUTINE FSILS_LS_FREE
 
          SUBROUTINE FSILS_BC_CREATE(lhs, faIn, nNo, dof, BC_type,       &
-     &         gNodes, Val)
+     &         gNodes, Val, virtual)
             INCLUDE "FSILS_STRUCT.h"
             TYPE(FSILS_lhsType), INTENT(INOUT) :: lhs
             INTEGER(KIND=LSIP), INTENT(IN) :: faIn, nNo, dof
             INTEGER(KIND=LSIP), INTENT(IN) :: BC_type
             INTEGER(KIND=LSIP), INTENT(IN) :: gNodes(nNo)
             REAL(KIND=LSRP), INTENT(IN), OPTIONAL :: Val(dof,nNo)
+            LOGICAL, INTENT(IN), OPTIONAL :: virtual
+            
          END SUBROUTINE FSILS_BC_CREATE
          SUBROUTINE external_BC_CREATE(lhs, faIn, nNo, dof, BC_type,    &
      &         gNodes, Val)
