### 2022-06-10 1dc982d
Updated shell formulation
- Accounts for hyperelastic constitutive models
- Implemented for incompressible materials only where plane stress is
satisfied analytically using static condensation
- Tested for both NURBS and triangular elements

TODO:

- Extend to compressible materials
- Add post-processing quantities for shell elements (stresses, strains, jacobian)
- Shells for 4-node or higher order elements (tri6, quad8, quad9)
