### 2022-06-11 a5da59f
Updated shell formulation for hyperelastic materials
- Compressible material is now implemented and working fine..

TODO:
- Post-processing quantities for shells
- Validate shell formulation using benchmark convergence tests
- Shells for higher order elements (quad8, quad9, tri6)
