+++ Code/Source/svFSI/MATMODELS.f
@@ -1134,7 +1134,8 @@ c     2      (EXP(stM%khs*Ess) + EXP(-stM%khs*Ess) + 2.0_RKIND)
       REAL(KIND=RKIND), INTENT(IN) :: gg_0(2,2), gg_x(2,2)
       REAL(KIND=RKIND), INTENT(OUT) :: Sml(3), Dml(3,3)
 
-      REAL(KIND=8) :: Jg2i, mu, gi_0(2,2), gi_x(2,2), S(2,2),CC(2,2,2,2)
+      REAL(KIND=RKIND) :: Jg2i, mu, gi_0(2,2), gi_x(2,2), S(2,2),
+     2   CC(2,2,2,2)
       TYPE(stModelType) :: stM
 
       Sml  = 0._RKIND
@@ -1143,7 +1144,7 @@ c     2      (EXP(stM%khs*Ess) + EXP(-stM%khs*Ess) + 2.0_RKIND)
 !     Some preliminaries
       stM  = lDmn%stM
 
-!     Contravariants in shell continuum
+!     Inverse of metric coefficients in shell continuum
       gi_0 = MAT_INV(gg_0, 2)
       gi_x = MAT_INV(gg_x, 2)
 
@@ -1196,13 +1197,115 @@ c     2      (EXP(stM%khs*Ess) + EXP(-stM%khs*Ess) + 2.0_RKIND)
       REAL(KIND=RKIND), INTENT(IN) :: gg_0(2,2), gg_x(2,2)
       REAL(KIND=RKIND), INTENT(OUT) :: Sml(3), Dml(3,3)
 
+      INTEGER(KIND=IKIND), PARAMETER :: MAXITR = 20
+      REAL(KIND=RKIND), PARAMETER :: ATOL = 1E-10
+
+      INTEGER(KIND=IKIND) :: itr, i, j, k, l
+      REAL(KIND=RKIND) :: Jg2, J2, J23, f13, f23, trC3, C33, kap, mu,
+     2   pJ, plJ, gi_x(2,2), gi_0(3,3), Ci(3,3), S(3,3), CC(3,3,3,3)
       TYPE(stModelType) :: stM
 
       Sml  = 0._RKIND
       Dml  = 0._RKIND
 
+!     Initialize tensor operations
+      CALL TEN_INIT(3)
+
 !     Some preliminaries
       stM  = lDmn%stM
+      kap  = stM%Kpen
+      mu   = 2._RKIND * stM%C10
+      f13  = 1._RKIND / 3._RKIND
+      f23  = 2._RKIND / 3._RKIND
+
+!     Inverse of metric coefficients in shell continuum
+      gi_x = MAT_INV(gg_x, 2)
+
+      gi_0 = 0._RKIND
+      gi_0(1:2,1:2) = MAT_INV(gg_0, 2)
+      gi_0(3,3) = 1._RKIND
+
+!     Ratio of inplane Jacobian determinant squared
+      Jg2 = MAT_DET(gg_x, 2) / MAT_DET(gg_0, 2)
+
+!     Begin Newton iterations to satisfy plane-stress condition.
+!     The objective is to find C33 that satisfies S33 = 0.
+      itr = 0
+      C33 = 1._RKIND
+      DO
+         itr  = itr + 1
+
+!        Trace (C)
+         trC3 = (gg_x(1,1)*gi_0(1,1) + gg_x(1,2)*gi_0(1,2)
+     2        +  gg_x(2,1)*gi_0(2,1) + gg_x(2,2)*gi_0(2,2) + C33)*f13
+
+!        Jacobian-related quantities
+         J2  = Jg2*C33
+         J23 = J2**(-f13)
+
+!        Inverse of curvilinear Cauchy-Green deformation tensor
+         Ci(:,:) = 0._RKIND
[…]
