### 2022-06-22 #4118 @axel-loewe
I can confirm that the Docker container is working and reproduces the example that I tried (04/06) :+1:

08/03 did start but seems to be a rather big simulation, did not complete within a couple of hours.

I'll proceed with the review when the responses to all comments are ready.
