### 2022-07-06 7bf1a6e
fixing some bugs
Current code produces segfault on Sherlock. Not completely fixed, but see changes.
