### 2022-07-10 222762a
BUG: fix verbose=0 will lead to solver crash due to missing result folder.
