+++ README.md
@@ -117,7 +117,7 @@ Below, we provide a list of the available choice of constitutive models for diff
    | cepModel\_TTP  | tenTusscher-Panfilov model       | "TTP", "tenTusscher-Panfilov" |
 
 ## Documentation
-More details can be found on the [**SimCardio**](http://simvascular.github.io/docsSimCardio.html) page, and direct links to the documentation for different funtionalities are provided here:
+More details can be found on the [**SimCardio**](http://simvascular.github.io/docsSimCardio.html) page, and direct links to the documentation for different functionalities are provided here:
 - Fluid-Structure Interaction (FSI): http://simvascular.github.io/docsSimCardio.html#fluid-structure-interaction
 - Cardiac electrophysiology modeling: http://simvascular.github.io/docsSimCardio.html#cep-modeling
 - Cardiac mechanics modeling:  http://simvascular.github.io/docsSimCardio.html#mechanics-modeling
@@ -128,8 +128,8 @@ More details can be found on the [**SimCardio**](http://simvascular.github.io/do
 - SimVascular group uploads hands-on tutorials to our [YouTube](https://www.youtube.com/channel/UCT61XgTRqpfb39Hyio9IqGQ) channel periodically. Here are some for `svFSI`:
   - Fluid-Structure Interaction (FSI): https://www.youtube.com/watch?v=QIpyThIAD7k
   - Cardiac electrophysiology modeling: https://www.youtube.com/watch?v=TCK3SmGwBa8
-  - Cardiac mechancis modeling: https://www.youtube.com/watch?v=Jm3VSi6Aci8
-- We also maintain a large collection of examples that showcase different funtionalities of `svFSI`. You can find them here: https://github.com/SimVascular/svFSI-Tests. Each case includes a README file that explains the problem in hand and some key aspects of the software configuration.
+  - Cardiac mechanics modeling: https://www.youtube.com/watch?v=Jm3VSi6Aci8
+- We also maintain a large collection of examples that showcase different functionalities of `svFSI`. You can find them here: https://github.com/SimVascular/svFSI-Tests. Each case includes a README file that explains the problem in hand and some key aspects of the software configuration.
 
 ## Pre/Post Processing Tool
 We are also maintaining a collection of useful pre and post processing tools that are compatible with `svFSI`:
