### 2022-07-10 1a5cc53
Reorganizing parallelization code for virtual face
Reorganizing parallelization code for virtual face handling into functions.

Specifically, in ALLFUN.f, moved communication code for virtual face integration into GatherMasterS and GatherMasterV, which are functions that collect nodal values (like nodal position) onto master, so that master can compute things for a virtual face element, such as normal vector or integration.

Also, in Distribute.f, moved handling for virtual face in PARTFACE() into a function PARTFACEV(), which sets up the parallel data structures for a virtual face.
