+++ Code/Source/svFSI/ALLFUN.f
@@ -90,6 +90,139 @@
       END INTERFACE
 
       CONTAINS
+
+!--------------------------------------------------------------------
+!     This routine gathers a nodal vector from a locals arrays across all procs 
+!     and sets a local vector to that value on master, and to zero on slaves
+!     
+!     In IntegV, a integral over a virtual face is performed by master
+!     alone, which must gather virtual face nodal values from other
+!     procs if master does not own the desired node. This function gathers
+!     all nodal values to master by performing the MPI communication
+!     among procs. 
+!
+!     ARGS:
+!     s(nsd, tnNo): a 2D array of values at each node. Ex. position vector at each node
+!     Ac: Index of node belonging to this proc. If Ac = 0, then the desired
+!     node does not belong to this proc
+!     snode(nsd): The vector value at the desired node. At the end of this routine
+!     snode is the vector value at the desired node on master, and zero on all slaves.
+!
+      SUBROUTINE GatherMasterV(s, Ac, snode)
+      USE COMMOD
+      IMPLICIT NONE
+      ! Array containing values at all nodes. This has length tnNo, which
+      ! is the number of nodes that belong to this proc
+      REAL(KIND=RKIND), INTENT(IN) :: s(:,:)
+      ! The proc local index of the node to query. If Ac = 0, then this
+      ! proc doesn't own the node to query, so we need to communicate
+      INTEGER(KIND=IKIND), INTENT(IN) :: Ac
+      ! Vector containing s at a node
+      REAL(KIND=RKIND), INTENT(OUT) :: snode(nsd)
+   
+      ! Allocate useful variables
+      INTEGER(KIND=IKIND) p, ierr
+      
+      ! Temporary array containing snode from all procs
+      REAL(KIND=RKIND), ALLOCATABLE :: sgather(:,:) 
+      ALLOCATE(sgather(nsd, cm%np()))
+
+      IF (Ac .EQ. 0) THEN ! This proc doesn't own this node
+         snode = 0._RKIND
+      ELSE ! This proc does own this node
+         snode = s(:,Ac) ! Get nodal function value to use
+      END IF
+!     Communicate the true snode value among all procs. We use the 
+!     gather operation to get snode from all procs onto master
+      ! First, allocate space for snode values. The size of sgather
+      ! vector is the number of procs being used.
+      sgather = 0._RKIND
+      CALL MPI_GATHER(snode, nsd, mpreal, sgather, nsd,
+     2            mpreal, master, cm%com(), ierr)
+!     If master, go through sgather and search for first non-zero
+!     value.
+      IF (cm%mas()) THEN
+         DO p=1, cm%np()
+            IF (.NOT. ALL(sgather(:,p) .EQ. 0)) THEN
+               snode(:) = sgather(:,p)
+               EXIT
+            END IF
+         END DO
+      ELSE ! If not master, set snode to 0 so they don't contribute to integral
+         snode = 0._RKIND 
+      END IF 
+
+!     Clean up
+      DEALLOCATE(sgather)
+
+      RETURN
+      END SUBROUTINE GatherMasterV
+
+!--------------------------------------------------------------------
+!     This routine gathers a nodal scalar from a locals arrays across all procs 
+!     and sets a local variable to that value on master, and to zero on slaves.
+!     
+!     In IntegS, a integral over a virtual face is performed by master
+!     alone, which must gather virtual face nodal values from other
+!     procs if master does not own the desired node. This function gathers
+!     all nodal values to master by performing the MPI communication
+!     among procs. 
+!
+!     ARGS:
[…]
