### 2022-07-12 9fe9ac1
Update README to reflect recent changes on simvascualr website.
