### 2022-07-12 b176ff4
cleanup and minor bug fixes
Just commenting or removing print statements, and copying some code snippets to similar parts of svFSI, which I forgot to do first time around
