+++ Code/Source/svFSILS/PRECOND.f
@@ -132,7 +132,7 @@
 !     Set valM(i,a) = val(i,a) * W(i,Ac)
       DO faIn=1, lhs%nFaces
          IF (lhs%face(faIn)%coupledFlag) THEN
-            PRINT*, "Under coupledFlag faIn", faIn
+!            PRINT*, "Under coupledFlag faIn", faIn
             DO a=1, lhs%face(faIn)%nNo
                Ac = lhs%face(faIn)%glob(a)
                DO i=1, MIN(lhs%face(faIn)%dof,dof)
