+++ Code/Source/svFSI/READFILES.f
@@ -888,6 +888,20 @@
          CALL READBC(lEq%bc(iBc), lPBC, lEq%phys)
       END DO
 
+!     If an LPN-coupled face has a cap, automatically create a BC
+!     for the cap face. This is necessary because we need svFSI to process
+!     the cap face as a coupled BC to add its contribution to the tangent
+      DO iBc=1, lEq%nBc
+         IF (BTEST(lEq%bc(iBc)%bType, bType_cpl)) THEN
+            IF (lEq%bc(iBc)%capFaceName .NE. "") THEN ! If coupled BC has a cap
+               ! Add a bc to lEq%bc(:) at the end for the cap, and add
+               ! cap face info to face being capped (capFaceName and
+               ! capFaceID fields)
+               CALL ADDCAPBC(lEq, iBc)
+            END IF
+         END IF 
+      END DO
+
 !     Initialize cplBC for RCR-type BC
       IF (ANY(BTEST(lEq%bc(:)%bType,bType_RCR))) THEN
          IF ((lEq%phys .NE. phys_fluid) .AND.
@@ -1621,6 +1635,9 @@
             err = "'Couple to cplBC' must be specified before"//
      2         " using Coupled BC"
          END IF
+
+!        AB 7/13/22: Read cap face name for this coupled BC
+         lPtr => list%get(lBc%capFaceName,"Cap")
       CASE ('Resistance')
          lBc%bType = IBSET(lBc%bType,bType_res)
          IF (.NOT.BTEST(lBc%bType,bType_Neu)) err = "Resistance"//
@@ -3153,4 +3170,61 @@ c     2         "can be applied for Neumann boundaries only"
 
       RETURN
       END SUBROUTINE READWALLPROPSFF
+!#######################################################################
+!     Adds a bc to lEq%bc(:) at the end for a cap. Copies most of bc info
+!     from lEq%bc(iBc), which corresponds to the surface being capped.
+!     Also, sets info about cap face in face being capped (capFaceName 
+!     and capFaceID fields)
+      SUBROUTINE ADDCAPBC(lEq, iBc)
+      USE COMMOD
+      USE ALLFUN
+      IMPLICIT NONE
+      TYPE(eqType), INTENT(INOUT) :: lEq
+      INTEGER(KIND=IKIND), INTENT(IN) :: iBc
+      TYPE(bcType), ALLOCATABLE :: oldBcs(:)
+      INTEGER(KIND=IKIND) jBc, iFa, iM
+
+!     We are adding a new BC for the cap, so we need to update the
+!     the relevant structures
+      ! Store old BCs
+      ALLOCATE(oldBcs(lEq%nBc))
+      oldBcs = lEq%bc
+      lEq%nBc = lEq%nBc + 1
+      DEALLOCATE(lEq%bc)
+      ALLOCATE(lEq%bc(lEq%nBc))
+
+      ! Copy old BCs to lEq%bc
+      DO jBc=1, lEq%nBc-1
+         lEq%bc(jBc) = oldBcs(jBc)
+      END DO
+
+      ! Add on new BC for cap. Copy BC information from surface being capped
+      ! This surface corresponds to iBc
+      lEq%bc(lEq%nBc) = lEq%bc(iBc)
+
+      ! Correct some values in cap BC (corresponding to nBc)
+      cplBC%nFa = cplBC%nFa + 1
+      lEq%bc(lEq%nBc)%cplBcPtr =cplBC%nFa
+      CALL FINDFACE(lEq%bc(iBc)%capFaceName, 
[…]
