### 2022-07-14 b4d34c0
fix memory bugs, now runs on Sherlock
