### 2022-07-15 #4118 @CZHU20
Dear all, thank you very much for your patience. We have finally completed all revisions. Here are our responses to @axel-loewe 's comments.

> Zhu et al. present svFSI, a simulation software for cardiac physiology. It covers multiple cardiac functions (physics): solid mechanics, hemodynamics and apparently cardiac electrophysiology. The effort of developing a software for this challenging coupled problem and making if publicly available is appreciated.
>
We thank the reviewer for the thorough evaluation of the software, and we have used these comments to improve our software.

> > License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?
>
> While there is no `LICENSE` file, there is `Copyright-Simvasscular.txt` which holds a license. Renaming might help to make this information more easily findable. The text seems to be the MIT license, I recommend mentioning this explicitly and also adding this information to the metadate of the GitHub project.
>
Per the reviewer’s recommendation, we have renamed Copyright-Simvasscular.txt to Licence.txt.
Since multiple parities were involved in the development of this software at different phases, we have to put our license file into the metadata of GitHub to reflect that. We have modified the [README](https://github.com/SimVascular/svFSI#introduction) to clearly state that the main body of svFSI follows MIT style license

> > Contribution and authorship: Has the submitting author (@CZHU20) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
>
> From the commit history, this does not seem to be the case (3 minor commits only). I am aware that the public commits may not be representative of the entire history of the project. Please comment on it.
>
Chi Zhu and Vijay Vedula share co-first authorship of the submission. Due to the submitting author’s account settings, some of the commits from CZHU20 are not automatically registered by GitHub, e.g. commit 1f24ab5fe513a1aa48b962e64e9718db6bb1d046 on Jun 30, 2021. Also, many changes made by the submitting author were added to svFSI through methods like diff-patch.

> > Installation: Does installation proceed as outlined in the documentation?
>
> I did not manage to compile the code with clang 13.0 (Undefined identifiers in METISLib). Do you see an option to increase compatibility with other compilers or to provide a container? In the current form, I cannot evaluate svFSI's functionality.
>
We are sorry that the reviewer was not able to compile the code. We have followed the reviewer’s suggestion and created a dockerfile to automate the build process. It is now included in the svFSI repository, and you can follow the instructions here to build and test it.
https://github.com/SimVascular/svFSI/tree/JOSS_paper/Docker

> > Functionality: Have the functional claims of the software been confirmed?
>
> svFSI is introduced as a multi-physics simulator including electrophysiology, tissue mechanics and blood dynamics. Unfortunately, I could not find any information on electrophysiology in the paper or the documentation (http://simvascular.github.io/docssvFSI.html). As such, I suggest to either document these features or to remove these aspects from the paper. Later on I found this link at the bottom of the README file: http://simvascular.github.io/docsSimCardio.html#cep-modeling It's not entirely clear to me how SimCardio relates to svFSI and I strongly recommend to provide a comprehensive overview of all relevant documentation for svFSI at a central place.
>
We appreciate the reviewer’s effort to locate these documents. The [SimCardio project](http://simvascular.github.io/docsSimCardio.html) aims to develop a complete software suite that can cover all aspects of cardiac modeling from image segmentation to computational modeling. svFSI is a major part of SimCardio as it performs multi-physics cardiac simulations.
Following changes have been made in [svFSI/README.md](https://github.com/SimVascular/svFSI#introduction) to better explain the scope of svFSI:
-	In the Introduction section, we added, “It[svFSI] is a major component of the ongoing SimVascular [**SimCardio**](http://simvascular.github.io/docsSimCardio.html) project that aims to provide the complete pipeline for cardiac modeling, from image segmentation to computational modeling.”
-	Following the reviewer’s suggestion, we have also overhauled the [svFSI website](http://simvascular.github.io/docssvFSI.html) where all relevant documentation for svFSI is now hosted. In the [Main features](http://simvascular.github.io/docssvFSI.html#feature) section, we also provide direct links to documentations of [fluid](http://simvascular.github.io/docssvFSI-Fluid.html), [structure](http://simvascular.github.io/docssvFSI-Structure.html), [electrophysiology](http://simvascular.github.io/docssvFSI-CEP.html) and [fluid-structure interaction](http://simvascular.github.io/docssvFSI-FSI.html).
-	Meanwhile, we have modified the Documentation and Tutorial section in [svFSI/README.md](https://github.com/SimVascular/svFSI#documentation) to better direct users to the right resource.

> > Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
>
> The website lists them but there is no automated way of installing them. Later on, I discovered more detailed information in the `INSTALL-DEPS.md` file in the repository. Would be good to sync this information with the web page documentation. Given the high number of dependencies and very specific versions being referred to in the documentation, I encourage you to consider providing a containerized version.
>
We have followed the reviewer’s suggestion and created a dockerfile to automate the build process. Please follow the instruction here to build the Docker image and test the software.
https://github.com/SimVascular/svFSI/tree/JOSS_paper/Docker

Clear build/installation instruction is also provided on the [website](http://simvascular.github.io/docssvFSI.html#install).

> > State of the field: Do the authors describe how this software compares to other commonly-used packages?
>
> > Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
>
> As coupling multiple physics is a complex problem and also multiplies the number of user-chosen parameters, types of boundary conditions, pre- and postprocessing steps etc., I assume that svFSI probably provides only a subset of features of established single-physics simulators. I believe it would be important for potential new users to learn about which features are supported. The brief paper remains rather vagues and does not mention the concrete scope. The documentation I found on the website is a tutorial for setting up a fluid-structure simulation (which I believe will be valuable for new users). However, I could not find a comprehensive user manual. The `Documentation` folder only contains a Doxygen configuration file. I feel that a proper user documentation is a prerequisite for svFSI to be really useful beyond the group of developers.
>
We agree with the reviewer that it is important to properly document the features of svFSI. To achieve this goal, we have first overhauled our [website](http://simvascular.github.io/docssvFSI.html)  to properly present the scope, features, theory, implementation and options of svFSI as well as its major components (fluid, structure, fsi, electrophysiology).

Furthermore, we have prepared a comprehensive list of examples to demonstrate almost all functionalities of svFSI, including physics solved, parameter selections, types of boundary conditions etc. From svFSI repository or website, users will be directed to https://github.com/SimVascular/svFSI-Tests. In each case, we have documented the problem it solves and key aspects of the software configuration. All cases have been tested and are ready to run by the user. Here are some examples:
- [04-fluid/04-3D0D-coupling-BC](https://github.com/SimVascular/svFSI-Tests/tree/master/04-fluid/04-3D0D-coupling-BC): generic boundary condition that couples 0D model to 3D flow
- [07-fsi/cmm/02-vein-graft](https://github.com/SimVascular/svFSI-Tests/tree/master/07-fsi/cmm/02-vein-graft): FSI simulation in a stenosed vein graft

We have established a new repository [svFSI-Tools](https://github.com/SimVascular/svFSI-Tools) to host all the pre/post-processing tools and added a new section Pre/Post Processing Tool to [svFSI/README.md](https://github.com/SimVascular/svFSI#prepost-processing-tool).

> > Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
>
> Could not find such information. There are no CI pipelines implemented as GitHub actions. Cannot judge on external CI.
>
We have significantly enhanced the [svFSI-Tests](https://github.com/SimVascular/svFSI-Tests) by including established simulation results in most of the examples (highlighted in bold in svFSI-Tests README), which can be used by the users to manually verify the functionality of the software.
The following sentence has been added to Section Features in svFSI/README.md:
“… the GitHub repository svFSI-Tests to get a detailed insight into the capability of the code. Also, most of the examples contain established simulation results, which users can use to verify the functionality of svFSI.”

> > Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
>
> Could not find such information in the repository or the svFSI part of the SimVascular webpage.
>
We have added new sections in both [svFSI README]( https://github.com/SimVascular/svFSI#contribute-to-svfsi) and [svFSI website]( http://simvascular.github.io/docssvFSI.html#resource) to provide clear guideline for these scenarios.

> > References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?
>
> Line 27: Please change the reference to openCARP to this one: https://dx.doi.org/10.1016/j.cmpb.2021.106223
>
The suggested change has been made.

> * Is the `Readme.md` file required? Seems it does not hold relevant information for users/developers and might be confused with `README.md`. Will also cause trouble on non-case-sensitive file systems.

This problem has been fixed by renaming `Readme.md` to `README_JOSS.md`.

Thank you all again.
