### 2022-07-15 #4118 @axel-loewe
I suggest mentioning the Docker container also in the [INSTALL.md file](https://github.com/SimVascular/svFSI/blob/JOSS_paper/INSTALL.md)

The authors markedly improved the documentation. For the future it would be great if there was also a complete user manual / documentation of parameters besides the examples.
