### 2022-07-15 b363108
Update INSTALL.md to include docker information.
