### 2022-07-15 #4118 @CZHU20
Hi @axel-loewe, thanks for the quick feedback. We have added the following sentence to [INSTALL.md](https://github.com/SimVascular/svFSI/blob/JOSS_paper/INSTALL.md) to include Docker container information.

>If you are interested in testing the new features in the most recent commit, we recommend you to build and run svFSI in [Docker container](https://github.com/SimVascular/svFSI/blob/JOSS_paper/Docker/README.md). Otherwise, please follow the instructions below to build svFSI on Ubuntu and MacOS.

A complete user manual and developer guide are definitely planned for the future. But users are always welcome to reach out to us with any usage/development questions.

Meanwhile, please let me know if you have any other comments. Thank you.
