### 2022-07-28 #4118 @axel-loewe
> @axel-loewe thanks for your help here. If you can, can you comment on @CZHU20 's response ☝️ ? Let me know, since you appear to have all boxes ticked, if you recommend this work for acceptance in JOSS or if there are still open issues. Thanks.

Yes, I recommend this work for publication!
