### 2022-07-29 #4118 @JaroslavHron
Hi, I have looked at the current state and it looks good - nice work.

- The docker image makes it easy to run the tests. I was able to compile/run the code without any problem (ubuntu, gcc).
- May be I would recommend to mention/cite in the text of the paper that for parallel solvers the trilinos library is used for the parallel linear solvers/preconditioners (or required/recommended for some of the examples - https://github.com/SimVascular/svFSI-Tests/tree/master/07-fsi/ale/03-pipe_3D)
