### 2022-07-29 eebc246
Add Trilinos citation to JOSS paper.
