### 2022-07-29 #4118 @CZHU20
@JaroslavHron Thank you very much for your comment.

I have revised the paper by adding the following sentence.

"It [svFS] also provides several built-in iterative linear solvers and preconditioners, along with many other options through Trilinos(Heroux et al., 2005)."
