+++ Code/Source/svFSI/READFILES.f
@@ -668,7 +668,7 @@
             propL(5,1) = damping
             propL(6,1) = f_x
             propL(7,1) = f_y
-            IF (nsd .EQ. 3) propL(8,1) = f_z
+            propL(8,1) = f_z
             IF (.NOT.cmmVarWall) THEN
                propL(9,1 ) = shell_thickness
                propL(10,1) = elasticity_modulus
@@ -688,13 +688,14 @@
             outPuts(11) = out_divergence
             outPuts(12) = out_acceleration
          ELSE
-            propL(1,1) = poisson_ratio
-            propL(2,1) = f_x
-            propL(3,1) = f_y
-            IF (nsd .EQ. 3) propL(4,1) = f_z
+            propL(1,1) = solid_density
+            propL(2,1) = poisson_ratio
+            propL(3,1) = f_x
+            propL(4,1) = f_y
+            propL(5,1) = f_z
             IF (.NOT.cmmVarWall) THEN
-               propL(5,1) = shell_thickness
-               propL(6,1) = elasticity_modulus
+               propL(6,1) = shell_thickness
+               propL(7,1) = elasticity_modulus
             END IF
 
             IF (pstEq) THEN
@@ -708,7 +709,6 @@
          END IF
 
          CALL READDOMAIN(lEq, propL, list)
-         IF (cmmInit) lEq%dmn(:)%prop(solid_density) = 0._RKIND
 
          CALL READLS(lSolver_GMRES, lEq, list)
 
