### 2022-08-03 5c44f40
Updated excitation-contraction coupling for electromechanics
- Both active stress and active strain approaches are verified
to be consistent with single node tests
- Added non-CEP based uniform activation options
- Added references for ionic activation models and coupling
- Some minor bugs were fixed

TODO:

- Additional tests need to be done to verify excitation-contraction
coupling
- Couple struct/ustruct with LPN
- Allow Purkinje deformation
- Perform tests on univentricle and biventricle models for full cycle
electromechanics
