+++ Code/Source/svFSI/READFILES.f
@@ -96,6 +96,8 @@
          shlEq        = .FALSE.
          pstEq        = .FALSE.
          sstEq        = .FALSE.
+         cepEq        = .FALSE.
+         ecCpld       = .FALSE.
          ibFlag       = .FALSE.
 
          i = IARGC()
@@ -274,35 +276,72 @@
          END IF
       END DO
 
-      IF (cem%cpld) THEN
-         IF (nEq .EQ. 1) err = "Min equations (2) not solved for"//
-     2      " electro-mechanics coupling"
-         i = 0
-
-         DO iEq=1, nEq
-            IF (eq(iEq)%phys .EQ. phys_CEP .OR.
-     2          eq(iEq)%phys .EQ. phys_struct .OR.
-     3          eq(iEq)%phys .EQ. phys_ustruct) i = i + 1
-         END DO
-         IF (i .NE. 2) err = "Both electrophysiology and struct have"//
-     2      " to be solved for electro-mechanics"
-
-         IF (cem%aStress .AND. cem%aStrain) err = "Cannot set "//
-     2      "both active strain and active stress coupling"
-
-         IF (cem%aStrain) THEN
-            IF (nsd .NE. 3) err = "Active strain coupling is allowed"//
-     2         " only for 3D bodies"
-            DO iEq=1, nEq
-               DO i=1, eq(iEq)%nDmn
-                  IF (eq(iEq)%dmn(i)%phys .NE. phys_ustruct .AND.
-     2                eq(iEq)%dmn(i)%phys .NE. phys_struct) CYCLE
-                  IF (eq(iEq)%dmn(i)%stM%isoType .NE. stIso_HO_d) err =
-     2               "Active strain is allowed with Holzapfel-Ogden "//
-     3               "passive constitutive model only"
+!     Check for inconsistencies in excitation-contraction model inputs
+!     for electromechanics simulations
+      IF (ecCpld) THEN
+         IF (nEq .EQ. 1) THEN
+!           Only struct/ustruct is allowed when nEq = 1
+            IF (eq(1)%phys.NE.phys_struct .AND.
+     2          eq(1)%phys.NE.phys_ustruct) THEN
+               err = " Min equations not solved for excitation-"//
+     2            "contraction coupling"
+            ELSE
+!              Only decoupled EC-coupling is allowed for pure struct
+!              and ustruct simulations
+               DO i=1, eq(1)%nDmn
+                  IF (eq(1)%dmn(i)%ec%caCpld) THEN
+                     err = " Invalid excitation-contraction coupling"
+                  END IF
                END DO
-            END DO
+            END IF
+         ELSE
+!           One of the equations solved must be `cep'
+            IF (.NOT.cepEq) THEN
+               err = " Electrophysiology equations should be solved"//
+     2            " for excitation-contraction coupling"
+            END IF
          END IF
+
+!        Check for inconsistencies in domain inputs with EC coupling
+         DO iEq=1, nEq
+            DO i=1, eq(iEq)%nDmn
+               flag = (eq(iEq)%dmn(i)%ec%astrain .OR.
+     2                 eq(iEq)%dmn(i)%ec%astress) .AND.
+     3                eq(iEq)%dmn(i)%cep%cepType .EQ. cepModel_FN
+               IF (flag) THEN
+                  err = " Excitation-contraction coupling is not"//
+     2               " allowed for Fitzhugh-Nagumo model"
+               END IF
+
[…]
