### 2022-08-03 #4118 @Kevin-Mattheus-Moerman
Hi @chennachaos I removed the one on top since you created the new checklist type above in a comment: https://github.com/openjournals/joss-reviews/issues/4118#issuecomment-1099385011

Let me know if you can use that, if not I'll put the one on top back.
