### 2022-08-08 #4118 @chennachaos
Hi @CZHU20,
I have the following comments regarding the software paper.

1.) Literature on the topic from Prof. Alfio Quarteroni's group, e.g. the iHEART project and the associated software (https://arxiv.org/pdf/2201.03303.pdf), is missing. Consider adding a couple of relevant references from this group.

2.) If possible, remove/merge the duplicates *equal contributions to the paper* and "Corresponding author" in the authors' details.
