+++ paper.bib
@@ -187,3 +187,39 @@ keywords = {Blood flow, Hemodynamics, Mass transport, Endothelial cells},
   publisher={ACM New York, NY, USA},
   url={https://doi.org/10.1145/1089014.1089021}
 }
+
+@article{quateroni_a,
+  doi = {10.48550/ARXIV.2201.03303},
+  
+  url = {https://arxiv.org/abs/2201.03303},
+  
+  author = {Africa, Pasquale C. and Piersanti, Roberto and Fedele, Marco and Dede', Luca and Quarteroni, Alfio},
+  
+  keywords = {Mathematical Software (cs.MS), Distributed, Parallel, and Cluster Computing (cs.DC), Numerical Analysis (math.NA), FOS: Computer and information sciences, FOS: Computer and information sciences, FOS: Mathematics, FOS: Mathematics, G.4; G.1; J.3, 68-04, 68N30 (Primary), 35-04, 65-04, 65M60, 65N30, 65Y05, 92-04, 92C50 (Secondary)},
+  
+  title = {life$^x$ - heart module: a high-performance simulator for the cardiac function. Package 1: Fiber generation},
+  
+  publisher = {arXiv},
+  
+  year = {2022},
+  
+  copyright = {Creative Commons Attribution Non Commercial Share Alike 4.0 International}
+}
+
+
+@article{quateroni_b,
+  doi = {10.48550/ARXIV.2207.12460},
+  url = {https://arxiv.org/abs/2207.12460},
+  
+  author = {Fedele, Marco and Piersanti, Roberto and Regazzoni, Francesco and Salvador, Matteo and Africa, Pasquale Claudio and Bucelli, Michele and Zingaro, Alberto and Dede', Luca and Quarteroni, Alfio},
+  
+  keywords = {Numerical Analysis (math.NA), Computational Engineering, Finance, and Science (cs.CE), Distributed, Parallel, and Cluster Computing (cs.DC), Medical Physics (physics.med-ph), FOS: Mathematics, FOS: Mathematics, FOS: Computer and information sciences, FOS: Computer and information sciences, FOS: Physical sciences, FOS: Physical sciences},
+  
+  title = {A comprehensive and biophysically detailed computational model of the whole human heart electromechanics},
+  
+  publisher = {arXiv},
+  
+  year = {2022},
+  
+  copyright = {Creative Commons Attribution Non Commercial No Derivatives 4.0 International}
+}
