+++ paper.md
@@ -7,19 +7,23 @@ tags:
   - fluid-structure interaction
   - finite element method
 authors:
- - name: Chi Zhu^[equal contributions to the paper]
+ - name: Chi Zhu
    orcid: 0000-0002-1099-8893
-   affiliation: "1, 2" # (Multiple affiliations must be quoted)
- - name: Vijay Vedula^[equal contributions to the paper]
+   equal-contrib: true
+   affiliation: "1, 2"
+ - name: Vijay Vedula
    affiliation: 3
+   equal-contrib: true
  - name: Dave Parker
    affiliation: 4
  - name: Nathan Wilson
    affiliation: 5
- - name: Shawn Shadden^[Corresponding author]
+ - name: Shawn Shadden
    affiliation: 1
- - name: Alison Marsden^[Corresponding author]
+   corresponding: true
+ - name: Alison Marsden
    affiliation: 4
+   corresponding: true
 affiliations:
  - name: University of California, Berkeley
    index: 1
@@ -46,7 +50,7 @@ Heart disease is the number one cause of death in the US [@Xu2020]. Many efforts
 
 # Statement of need
 
-Accompanying the growing popularity of studying cardiac physiology with computational modeling, many open-source software tools that specialize in modeling one or two aspects of the multiphysics process in the heart have been developed. For example, `SimVascular` [@Updegrove2017] enables patient-specific blood flow modeling by providing a complete pipeline from medical image segmentation to simulation results. `FEBio` [@Maas2012] specializes in modeling large-deformation structure mechanics in biophysics with fluid-structure interaction (FSI) capability as well. `openCARP` [@Plank2021] focuses on modeling cardiac electrophysiology. There are other general-purpose open-source software such as `LifeV` [@Bertagna2017] and `FEniCS` [@Logg2012], that can be flexibly adapted to simulate different physics in the heart, but significant development effort may be required for this purpose.
+Accompanying the growing popularity of studying cardiac physiology with computational modeling, many open-source software tools that specialize in modeling one or two aspects of the multiphysics process in the heart have been developed. For example, `SimVascular` [@Updegrove2017] enables patient-specific blood flow modeling by providing a complete pipeline from medical image segmentation to simulation results. `FEBio` [@Maas2012] specializes in modeling large-deformation structure mechanics in biophysics with fluid-structure interaction (FSI) capability as well. `openCARP` [@Plank2021] focuses on modeling cardiac electrophysiology. There are other general-purpose open-source software such as `LifeV` [@Bertagna2017] and `FEniCS` [@Logg2012], that can be flexibly adapted to simulate different physics in the heart, but significant development effort may be required for this purpose. Recently, Quarteroni et al. have also been developing an open-source simulator for the cardiac function, and the fiber generation module [@quateroni_a] and electromechanics [@quateroni_b] have been announced so far.
 
[…]
