### 2022-08-09 #4118 @CZHU20
@chennachaos  Thank you very much for your comment. We have addressed both in the new commit:

1. We have added the following sentence to the manuscript:

> Recently, Quarteroni et al. have also been developing an open-source simulator for the cardiac function, and the fiber generation module (Africa et al., 2022) and electromechanics (Fedele et al., 2022) have been announced so far.

2. Proper ways to assign equal contribution and corresponding authors have been adopted.
```
 - name: Chi Zhu
   orcid: 0000-0002-1099-8893
   equal-contrib: true
   affiliation: "1, 2"
 - name: Vijay Vedula
   affiliation: 3
   equal-contrib: true
 - name: Dave Parker
   affiliation: 4
 - name: Nathan Wilson
   affiliation: 5
 - name: Shawn Shadden
   affiliation: 1
   corresponding: true
 - name: Alison Marsden
   affiliation: 4
   corresponding: true
```

Please let me know if you have other comments. Thank you.
