### 2022-08-09 #4118 @CZHU20
@Kevin-Mattheus-Moerman
I have followed the JOSS guideline to assign equal contribution and corresponding author. However, the generated pdf does not reflect both changes. I used the following docker command to generate the pdf
```
docker run --rm \
    --volume $PWD:/data \
    --user $(id -u):$(id -g) \
    --env JOURNAL=joss \
    openjournals/paperdraft
```
Is this because it's currently in draft state? Thank you.
