### 2022-08-09 2547858
Added feature to read multiple fiber directions from single file
Other cosmetic changes..
