+++ Code/Source/svFSI/READMSH.f
@@ -46,7 +46,7 @@
 
       CHARACTER, PARAMETER :: dSym(3) = (/"X","Y","Z"/)
 
-      LOGICAL :: flag
+      LOGICAL :: flag, fib_dir
       INTEGER(KIND=IKIND) :: i, j, iM, iFa, a, b, Ac, e, lDof, lnNo
       REAL(KIND=RKIND) :: maxX(nsd), minX(nsd), fibN(nsd), rtmp
       CHARACTER(LEN=stdL) :: ctmp, fExt
@@ -330,52 +330,95 @@
          END DO
       END IF
 
-!     Read fiber orientation
-      flag = .FALSE.
+!     Read fiber orientation. Here we will try to load fibers in
+!     multiple ways. First we will attempt to read fibers from a single
+!     file. The code will look for variables named as "FIB_DIR1",
+!     "FIB_DIR2", etc. in a single vtu file.
+!     If not found, the code will then search for multiple fiber files
+!     (vtu format) with fiber direction variable named as "FIB_DIR".
+!     Otherwise, we will search for a uniform fiber vector with the
+!     keyword "Fiber direction".
+      dbg = " Checking for any fiber directions"
       DO iM=1, nMsh
-         lPM => list%get(msh(iM)%name,"Add mesh",iM)
-         j = lPM%srch("Fiber direction file path")
-         IF (j .EQ. 0) j = lPM%srch("Fiber direction")
-         IF (j .NE. 0) THEN
-            flag = .TRUE.
-            EXIT
-         END IF
-      END DO
-      IF (flag) THEN
-         DO iM=1, nMsh
-            lPM => list%get(msh(iM)%name,"Add mesh",iM)
+         msh(iM)%nFn = 0
 
+!        First attempt to read fibers from a single vtu file.
+         fib_dir = .FALSE.
+         lPM => list%get(msh(iM)%name, "Add mesh", iM)
+
+!        Check if num fiber directions are provided by the user
+         lPtr => lPM%get(msh(iM)%nFn, "Number of fiber directions")
+
+!        If not, the code will try to compute nFn from the variable
+!        names in the vtu file
+         lPtr => lPM%get(ctmp, "Fiber directions file path")
+         IF (ASSOCIATED(lPtr)) THEN
+            IF (rmsh%isReqd) err = "Fiber directions read from"//
+     2         " a file is not allowed with remeshing"
+            CALL READFIBNFSF(msh(iM), ctmp)
+            IF (ALLOCATED(msh(iM)%fN)) fib_dir = .TRUE.
+         END IF
+
+!        If the fibers are not found, we will try searching for multiple
+!        fiber files
+         IF (.NOT.fib_dir) THEN
+            lPM => list%get(msh(iM)%name, "Add mesh", iM)
             msh(iM)%nFn = lPM%srch("Fiber direction file path")
+
             IF (msh(iM)%nFn .NE. 0) THEN
+               fib_dir = .TRUE.
                IF (rmsh%isReqd) err = "Fiber directions read from "//
      2            "file is not allowed with remeshing"
-               ALLOCATE(msh(iM)%fN(msh(iM)%nFn*nsd,msh(iM)%gnEl))
+               ALLOCATE(msh(iM)%fN(nsd*msh(iM)%nFn,msh(iM)%gnEl))
                msh(iM)%fN = 0._RKIND
+
                DO i=1, msh(iM)%nFn
-                  lPtr => lPM%get(cTmp,
[…]
