### 2022-08-15 ad7e92e
Bug fixed in I8fs invariant terms for stress and stiffness tensors
The bug affects decoupled-HO and modified-HO models only

Changed transmural coordinate variable to Phi_EPI
