### 2022-08-16 #4118 @chennachaos
Hi @CZHU20,
Some nice examples are provided in [svFSI-Tests](https://github.com/SimVascular/svFSI-Tests) to help the users verify the software. However, it looks like README files for some of the examples need to be updated. For example, all three ALE FSI examples point to the same YouTube tutorial (on pressure pulse inside the aorta).

I recommend checking all the examples thoroughly and updating them as necessary. I would also recommend adding comparisons of results, either as graphs or tables, to the corresponding README files so that the user knows straightway what to expect when the input files are executed, instead of navigating through lengthy journal papers.
