### 2022-08-25 40eddf4
Updates in excitation-contraction coupling
- Bug fixed while reading a constant fiber direction
- Some changes to the data structures for excitation-contraction coupling
- Added an option to prescribe fiber shortening from a file (both
steady/unsteady)
- Tested active-strain on a idealized cube by prescribing shortening
data from Rossi 2014 paper.

TODO: need to test electrophysiology-coupled excitation-contraction
(Rossi 2014 - BO model coupled with orthotropic active strain over a
cubical domain)
