+++ Code/Source/svFSI/DISTRIBUTE.f
@@ -579,20 +579,14 @@
             END IF
          END IF
 
-         IF (ecCpld) THEN
-            CALL cm%bcast(lEq%dmn(iDmn)%ec%caCpld)
-            CALL cm%bcast(lEq%dmn(iDmn)%ec%astress)
-            CALL cm%bcast(lEq%dmn(iDmn)%ec%astrain)
-            CALL cm%bcast(lEq%dmn(iDmn)%ec%asnType)
-            CALL cm%bcast(lEq%dmn(iDmn)%ec%odes%tIntType)
-            CALL cm%bcast(lEq%dmn(iDmn)%ec%fpar_in)
-            CALL cm%bcast(lEq%dmn(iDmn)%ec%dt)
-         END IF
-
          IF ((lEq%dmn(iDmn)%phys .EQ. phys_struct)  .OR.
      2       (lEq%dmn(iDmn)%phys .EQ. phys_ustruct) .OR.
      3       (lEq%dmn(iDmn)%phys .EQ. phys_shell)) THEN
             CALL DIST_MATCONSTS(lEq%dmn(iDmn)%stM)
+         END IF
+
+         IF (ecCpld) THEN
+            CALL DIST_ECMODEL(lEq%dmn(iDmn)%ec)
          END IF
 
          IF ((lEq%dmn(iDmn)%phys .EQ. phys_fluid)  .OR.
@@ -1034,7 +1028,7 @@
       END SUBROUTINE DISTBF
 !--------------------------------------------------------------------
 !     This subroutine distributes constants and parameters of the
-!     constitutive model to all processes
+!     structural constitutive model to all processes
       SUBROUTINE DIST_MATCONSTS(lStM)
       USE COMMOD
       USE ALLFUN
@@ -1087,7 +1081,51 @@
       END SUBROUTINE DIST_MATCONSTS
 !--------------------------------------------------------------------
 !     This subroutine distributes constants and parameters of the
-!     constitutive model to all processes
+!     excitation-contraction coupling model
+      SUBROUTINE DIST_ECMODEL(lEc)
+      USE COMMOD
+      USE ALLFUN
+      IMPLICIT NONE
+      TYPE(eccModelType), INTENT(INOUT) :: lEc
+
+      INTEGER(KIND=IKIND) i, j
+
+      CALL cm%bcast(lEc%caCpld)
+      CALL cm%bcast(lEc%astress)
+      CALL cm%bcast(lEc%astrain)
+      CALL cm%bcast(lEc%asnType)
+      CALL cm%bcast(lEc%k)
+      CALL cm%bcast(lEc%odes%tIntType)
+      CALL cm%bcast(lEc%fpar_in)
+      CALL cm%bcast(lEc%dt)
+      CALL cm%bcast(lEc%dType)
+      IF (BTEST(lEc%dType, bType_std)) THEN
+         CALL cm%bcast(lEc%Ya)
+      ELSE IF (BTEST(lEc%dType, bType_ustd)) THEN
+         CALL cm%bcast(lEc%Yat%lrmp)
+         CALL cm%bcast(lEc%Yat%d)
+         CALL cm%bcast(lEc%Yat%n)
+         j = lEc%Yat%d
+         i = lEc%Yat%n
+         IF (cm%slv()) THEN
+            ALLOCATE(lEc%Yat%qi(j))
+            ALLOCATE(lEc%Yat%qs(j))
+            ALLOCATE(lEc%Yat%r(j,i))
+            ALLOCATE(lEc%Yat%i(j,i))
+         END IF
+         CALL cm%bcast(lEc%Yat%ti)
+         CALL cm%bcast(lEc%Yat%T)
+         CALL cm%bcast(lEc%Yat%qi)
+         CALL cm%bcast(lEc%Yat%qs)
[…]
