### 2022-08-26 f249c33
Some key updates to excitation-contraction coupling
- Coupling between cep and struct/ustruct for electromechanics
- Tested for both prescribed and coupled electromechanics
- Both active stress and active strain are working fine
