+++ Code/Source/svFSI/CEPION.f
@@ -42,9 +42,9 @@
       USE ALLFUN
       IMPLICIT NONE
 
-      INTEGER(KIND=IKIND) a, iEq, iDmn, cPhys, dID, nX, nG
+      INTEGER(KIND=IKIND) a, iEq, iDmn, cPhys, dID, nX, nG, maxNx, maxNg
 
-      REAL(KIND=RKIND), ALLOCATABLE :: Xl(:), Xgl(:), sA(:), sF(:,:)
+      REAL(KIND=RKIND), ALLOCATABLE :: Xl(:,:), Xgl(:,:), sA(:), sF(:,:)
 
       DO iEq=1, nEq
          IF (eq(iEq)%phys .NE. phys_CEP) CYCLE
@@ -52,6 +52,25 @@
 !        Average cellular activation state variables at domain
 !        interfaces for multi-domain problem
          IF (ALLOCATED(dmnId)) THEN
+            maxNx = 0
+            maxNg = 0
+            DO iDmn=1, eq(iEq)%nDmn
+               nX = eq(iEq)%dmn(iDmn)%cep%nX
+               nG = eq(iEq)%dmn(iDmn)%cep%nG
+               IF (maxNx .LT. nX) maxNx = nX
+               IF (maxNg .LT. nG) maxNg = nG
+            END DO
+
+            ALLOCATE(Xl(maxNx,eq(iEq)%nDmn), Xgl(maxNg,eq(iEq)%nDmn))
+            Xl  = 0._RKIND
+            Xgl = 0._RKIND
+            DO iDmn=1, eq(iEq)%nDmn
+               nX = eq(iEq)%dmn(iDmn)%cep%nX
+               nG = eq(iEq)%dmn(iDmn)%cep%nG
+               CALL CEPINITL(eq(iEq)%dmn(iDmn)%cep, nX, nG,
+     2            Xl(1:nX,iDmn), Xgl(1:nG,iDmn))
+            END DO
+
             ALLOCATE(sA(tnNo), sF(nXion,tnNo))
             sA = 0._RKIND
             sF = 0._RKIND
@@ -62,14 +81,9 @@
                   dID   = eq(iEq)%dmn(iDmn)%Id
                   IF (cPhys.NE.phys_CEP .OR. .NOT.BTEST(dmnId(a),dID))
      2                CYCLE
-                  nX = eq(iEq)%dmn(iDmn)%cep%nX
-                  nG = eq(iEq)%dmn(iDmn)%cep%nG
-                  ALLOCATE(Xl(nX), Xgl(nG))
-                  CALL CEPINITL(eq(iEq)%dmn(iDmn)%cep, nX, nG, Xl, Xgl)
                   sA(a) = sA(a) + 1._RKIND
-                  sF(1:nX,a)  = sF(1:nX,a) + Xl(:)
-                  sF(nX+1:nX+nG,a) = sF(nX+1:nX+nG,a) + Xgl(:)
-                  DEALLOCATE(Xl, Xgl)
+                  sF(1:nX,a)  = sF(1:nX,a) + Xl(1:nX,iDmn)
+                  sF(nX+1:nX+nG,a) = sF(nX+1:nX+nG,a) + Xgl(1:nG,iDmn)
                END DO
             END DO
             CALL COMMU(sA)
@@ -78,18 +92,18 @@
                IF (.NOT.ISZERO(sA(a)))
      2            Xion(:,a) = sF(:,a)/sA(a)
             END DO
-            DEALLOCATE(sA, sF)
+            DEALLOCATE(Xl, Xgl, sA, sF)
          ELSE
+            nX = eq(iEq)%dmn(1)%cep%nX
[…]
