### 2022-08-27 63c000a
Some more updates to excitation-contraction coupling
- bug fixed while computing isotropic diffusion when coupled with mechancis
- added fiber shortening as an output parameter
- tested coupled electromechanics with active strain model on a simple cube

TODO: need to test coupled electromechanics on patient-specific LV with both
active stress and active strain models
