### 2022-08-31 9eac3b3
fix bugs in shell, add constitutive equs
