+++ Code/Source/svFSI/MATMODELS.f
@@ -1105,16 +1105,19 @@ c     2      (EXP(stM%khs*Ess) + EXP(-stM%khs*Ess) + 2.0_RKIND)
 !####################################################################
 !     Compute 2nd Piola-Kirchhoff stress and material stiffness tensors
 !     for incompressible shell elements
-      SUBROUTINE GETPK2CC_SHLi(lDmn, gg_0, gg_x, Sml, Dml)
+      SUBROUTINE GETPK2CC_SHLi(lDmn, nfd, fNa0, gg_0, gg_x, Sml, Dml)
       USE MATFUN
       USE COMMOD
       IMPLICIT NONE
       TYPE(dmnType), INTENT(IN) :: lDmn
-      REAL(KIND=RKIND), INTENT(IN) :: gg_0(2,2), gg_x(2,2)
+      INTEGER(KIND=IKIND), INTENT(IN) :: nfd
+      REAL(KIND=RKIND), INTENT(IN) :: gg_0(2,2), gg_x(2,2), fNa0(2,nfd)
       REAL(KIND=RKIND), INTENT(OUT) :: Sml(3), Dml(3,3)
 
-      REAL(KIND=RKIND) :: Jg2i, mu, gi_0(2,2), gi_x(2,2), S(2,2),
-     2   CC(2,2,2,2)
+      REAL(KIND=RKIND) :: Jg2i, I1, mu, gi_0(2,2), gi_x(2,2), S(2,2),
+     2   CC(2,2,2,2), SN(2,2), CCN(2,2,2,2), a, b, Inv4, Inv6, Inv8,
+     3   Eff, Ess, Efs, flM(2,2), c4f, c4s, dc4f, dc4s, d1, iFn,
+     4   fl(2,2,nfd), Hfs(2,2), EI1, I1T
       TYPE(stModelType) :: stM
 
       Sml  = 0._RKIND
@@ -1132,14 +1135,142 @@ c     2      (EXP(stM%khs*Ess) + EXP(-stM%khs*Ess) + 2.0_RKIND)
       IF (ISZERO(Jg2i)) err = " Divide by zero in-plane Jacobian"//
      2   " determinant"
       Jg2i = MAT_DET(gg_0, 2) / Jg2i
+      I1 = 0._RKIND
+      DO a = 1, 2
+        DO b = 1, 2
+            I1 = I1 + gi_0(a,b)*gg_x(a,b)
+        END DO
+      END DO
 
       SELECT CASE(stM%isoType)
       CASE (stIso_nHook)
          mu = 2._RKIND * stM%C10
          S  = mu*(gi_0 - Jg2i*gi_x)
 
-         CC = 2._RKIND*mu*(TEN_DYADPROD(gi_x, gi_x, 2) +
+         CC = 2._RKIND*mu*Jg2i*(TEN_DYADPROD(gi_x, gi_x, 2) +
      2                     TEN_SYMMPROD(gi_x, gi_x, 2))
+
+      CASE (stIso_MR)
+         SN = (gi_0 - Jg2i*gi_x)
+         CCN= 2._RKIND*Jg2i*(TEN_DYADPROD(gi_x, gi_x, 2) +
+     2                     TEN_SYMMPROD(gi_x, gi_x, 2))
+
+         S  = stM%C10*SN + stM%C01*Jg2i* (gi_0 - I1*gi_x)
+     2         + stM%C01/Jg2i*gi_x
+         CC = (stM%C10 + stM%C01*I1) * CCN - 2._RKIND*stM%C01 * 
+     2        Jg2i * (TEN_DYADPROD(gi_0, gi_x, 2) + 
+     3        TEN_DYADPROD(gi_x, gi_0, 2)) + 2._RKIND*stM%C01 /
[…]
