+++ Code/Source/svFSI/CONSTS.f
@@ -183,7 +183,8 @@
       INTEGER(KIND=IKIND), PARAMETER :: stIso_NA = 600, stIso_lin = 601,
      2   stIso_StVK = 602, stIso_mStVK = 603, stIso_nHook = 604,
      3   stIso_MR = 605, stIso_HGO_d = 606, stIso_HGO_ma = 607,
-     4   stIso_Gucci = 608, stIso_HO_d = 609, stIso_HO_ma = 610
+     4   stIso_Gucci = 608, stIso_HO_d = 609, stIso_HO_ma = 610,
+     5   stIso_L_Scks = 611
 !--------------------------------------------------------------------
 !     Type of constitutive model (volumetric) for structure eqn:
 !     Quadratic, Simo-Taylor91, Miehe94
