+++ Code/Source/svFSI/READFILES.f
@@ -2738,9 +2738,9 @@ c     2         "can be applied for Neumann boundaries only"
          lPtr => lSt%get(lDmn%stM%bfs, "bfs")
          lPtr => lSt%get(lDmn%stM%khs, "k")
 
-      CASE ("Lee-Sacks", "L-Scks")
-      ! Holzapefel and Ogden model for myocardium !
-         lDmn%stM%isoType = stIso_L_Scks
+      CASE ("Lee-Sacks", "L-Scks", "LS")
+      ! Lee-Sacks model for cardiac valves !
+         lDmn%stM%isoType = stIso_LS
          lPtr => lSt%get(lDmn%stM%a, "a")
          lPtr => lSt%get(lDmn%stM%a0, "a0")
          lPtr => lSt%get(lDmn%stM%b1, "b1")
@@ -2790,10 +2790,6 @@ c     2         "can be applied for Neumann boundaries only"
 
 !     Check for shell model
       IF (lDmn%phys .EQ. phys_shell) THEN
-         ! IF (lDmn%stM%isoType .NE. stIso_nHook) THEN
-         !    err = "Only Neo-Hookean model is allowed for shell elements"
-         ! END IF
-
 !        ST91 is the default and the only dilational penalty model for
 !        compressible shell elements. This is set to avoid any square-
 !        root evaulations of the Jacobian during Newton iterations for
