### 2022-08-31 a88216d
change compressible material for shell
