### 2022-08-31 7890639
Added parameter for cross-fiber stress contribution
The cross-fiber stress parameter adds stress along the sheet or
cross-fiber direction, relevant for cardiac mechanics.

Commit also include removing a few unused variables and declaring loop
variables as integers
