+++ Code/Source/svFSI/MOD.f
@@ -212,6 +212,8 @@
          REAL(KIND=8) :: g = 0._RKIND
 !        Unsteady time-dependent values
          TYPE(fcType) :: gt
+!        Cross-fiber (sheet) stress parameter
+         REAL(KIND=RKIND) :: eta_s = 0._RKIND
       END TYPE fibStrsType
 
 !     Structural domain type
@@ -278,6 +280,8 @@
 !          := activation force for active stress model
 !          := fiber contraction parameter for active strain model
          REAL(KIND=RKIND) :: Ya = 0._RKIND
+!        Cross-fiber stress component for active stress coupling
+         REAL(KIND=RKIND) :: eta_s = 0._RKIND
 !        Unsteady time-dependent values for prescribed fiber-shortening
          TYPE(fcType) :: Yat
       END TYPE eccModelType
