### 2022-09-01 #4118 @CZHU20
Hi @chennachaos
Thank you very much for your comments.
Regarding the YouTube tutorials, we currently only recorded some most common cases to demonstrate the main features of svFSI. Hence, all the README files of FSI cases pointed to this tutorial. I agree it's confusing, so we have added a clarification (see [here](https://github.com/SimVascular/svFSI-Tests/tree/master/07-fsi/ale/02-channel-block-flag_2D)):
> This video is for flow inside a modeled aorta, but the general idea is also applicable to this case.

We will gradually upload more tutorials to our channel.

We have also followed your suggestion and updated all of our README files to make them more self-explanatory. For those cases with a reference, we have added problem configuration and some results so that the readers can get all the information at once. Here are some examples:
Fluid: https://github.com/SimVascular/svFSI-Tests/tree/master/04-fluid/03-driven-cavity-2D
Structure: https://github.com/SimVascular/svFSI-Tests/tree/master/05-struct/02-LV-Guccione-passive
FSI: https://github.com/SimVascular/svFSI-Tests/tree/master/07-fsi/ale/03-pipe_3D
For some other cases without a reference, we have included some form of the results. For example:
Heat conduction: https://github.com/SimVascular/svFSI-Tests/tree/master/01-heat/01-diffusion-line-source
Electrophysiology: https://github.com/SimVascular/svFSI-Tests/tree/master/08-cep/05-Purkinje

I hope these changes address your concerns.
Thank you again for your time. Please let me know if you have other comments.
