### 2022-09-04 9ef49ee
Revert "Merge branch 'SimVascular:JOSS_paper' into virtual_face"
This reverts commit d742ba87781b99a57c973f0652ff59529f105a1e, reversing
changes made to e9ae2ef50ff20dc9e11f02ec22eb7c7043d57f2b.
