### 2022-09-11 #4118 @Kevin-Mattheus-Moerman
@CZHU20 since all reviewers are now happy we are ready to process this work for acceptance in JOSS. There a couple of steps remaining.

On the paper, I have read it and it looks good, I only have the following minor editorial comments:
- [ ] Please add the city and country to each of your affiliations. Do not use acronyms for countries, e.g. please spell out USA as United States of America.
- [ ] You may use `open source` rather than `open-source` throughout the paper.

Other points:
- [ ] Please can you conform what the version for the reviewed software should be? Are you still at `v1.0.0` or has the version tag changed?
- [ ] Please can you archive a copy of the software on [ZENODO](https://zenodo.org/). You can do this manually or use these automated steps: https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content
- [ ] Once the ZENODO archive is create please report back here with the ZENODO DOI.
- [ ] Please manually edit the ZENODO archive where needed to ensure that the title and authors match those on the JOSS paper title. Furthermore please check that the license listed there is correct.
