+++ paper.md
@@ -25,15 +25,15 @@ authors:
    affiliation: 4
    corresponding: true
 affiliations:
- - name: University of California, Berkeley
+ - name: University of California, Berkeley, United States of America
    index: 1
- - name: Peking University
+ - name: Peking University, Beijing, People's Republic of China
    index: 2
- - name: Columbia University
+ - name: Columbia University, New York City, United States of America
    index: 3
- - name: Stanford University
+ - name: Stanford University, Stanford, United States of America
    index: 4
- - name: University of California, Los Angeles
+ - name: University of California, Los Angeles, United States of America
    index: 5
 date: 20 January 2021
 bibliography: paper.bib
@@ -46,13 +46,13 @@ bibliography: paper.bib
 
 # Summary
 
-Heart disease is the number one cause of death in the US [@Xu2020]. Many efforts have been devoted to studying its progression, diagnosis, and treatment. During the past decade, computational modeling has made significant inroads into the research of heart disease. The heart is inherently a multiphysics system that includes electrophysiology, tissue mechanics, and blood dynamics. Its normal function starts with the propagation of electrical signals that trigger the active contraction of the heart muscle to pump blood into the circulatory system. Rooted in fundamental laws of physics such as the balance of mass, momentum, and energy, computational modeling has been instrumental in studying cardiac physiology such as left ventricular function [@Mittal2015], cardiac arrhythmia [@Trayanova2011], and blood flow in the cardiovascular system [@Arzani2018;@Grande2021]. `svFSI` is the first open-source software that specializes in enabling coupled electro-mechano-hemodynamic simulations of the heart.
+Heart disease is the number one cause of death in the US [@Xu2020]. Many efforts have been devoted to studying its progression, diagnosis, and treatment. During the past decade, computational modeling has made significant inroads into the research of heart disease. The heart is inherently a multiphysics system that includes electrophysiology, tissue mechanics, and blood dynamics. Its normal function starts with the propagation of electrical signals that trigger the active contraction of the heart muscle to pump blood into the circulatory system. Rooted in fundamental laws of physics such as the balance of mass, momentum, and energy, computational modeling has been instrumental in studying cardiac physiology such as left ventricular function [@Mittal2015], cardiac arrhythmia [@Trayanova2011], and blood flow in the cardiovascular system [@Arzani2018;@Grande2021]. `svFSI` is the first open source software that specializes in enabling coupled electro-mechano-hemodynamic simulations of the heart.
 
 # Statement of need
 
-Accompanying the growing popularity of studying cardiac physiology with computational modeling, many open-source software tools that specialize in modeling one or two aspects of the multiphysics process in the heart have been developed. For example, `SimVascular` [@Updegrove2017] enables patient-specific blood flow modeling by providing a complete pipeline from medical image segmentation to simulation results. `FEBio` [@Maas2012] specializes in modeling large-deformation structure mechanics in biophysics with fluid-structure interaction (FSI) capability as well. `openCARP` [@Plank2021] focuses on modeling cardiac electrophysiology. There are other general-purpose open-source software such as `LifeV` [@Bertagna2017] and `FEniCS` [@Logg2012], that can be flexibly adapted to simulate different physics in the heart, but significant development effort may be required for this purpose. Recently, Quarteroni et al. have also been developing an open-source simulator for the cardiac function, and the fiber generation module [@quateroni_a] and electromechanics [@quateroni_b] have been announced so far.
[…]
