### 2022-09-16 #4118 @CZHU20
@Kevin-Mattheus-Moerman Thank you.
I have fixed the affliction problem and used "open source" everywhere.

Regarding software version and ZENODO archive,  the development on the public repository has been frozen since the JOSS paper was submitted. After discussing with other authors, we intend to create a new tag for the current version and archive the same version in ZENODO. But we don't want to include the JOSS material (paper.md, paper.bib etc) as they are not part of the software. Can we merge changes (made according to reviewers' comments) in the JOSS branch to the main branch, and archive the main branch instead of the JOSS branch?

Thank you very much!
