### 2022-09-19 #4118 @Kevin-Mattheus-Moerman
@CZHU20 Yes, what you suggest regarding the archiving and leaving out the paper.md etc is fine.
