+++ Code/Source/svFSI/POST.f
@@ -66,7 +66,13 @@
             ALLOCATE(tmpV(1,msh(iM)%nNo), tmpVe(msh(iM)%nEl))
             tmpV  = 0._RKIND
             tmpVe = 0._RKIND
-            CALL TPOST(msh(iM), 1, tmpV, tmpVe, lD, lY, iEq, outGrp)
+
+            IF (msh(iM)%lShl) THEN
+               CALL SHLPOST(msh(iM), 1, tmpV, tmpVe, lD, iEq, outGrp)
+            ELSE
+               CALL TPOST(msh(iM), 1, tmpV, tmpVe, lD, lY, iEq, outGrp)
+            END  IF
+
             res  = 0._RKIND
             DO a=1, msh(iM)%nNo
                Ac = msh(iM)%gN(a)
@@ -898,6 +904,371 @@
 
       RETURN
       END SUBROUTINE TPOST
+!####################################################################
+!     Routine for post processing shell-based quantities
+      SUBROUTINE SHLPOST(lM, m, res, resE, lD, iEq, outGrp)
+      USE COMMOD
+      USE ALLFUN
+      USE MATFUN
+      IMPLICIT NONE
+      TYPE(mshType), INTENT(INOUT) :: lM
+      INTEGER(KIND=IKIND), INTENT(IN) :: m, iEq, outGrp
+      REAL(KIND=RKIND), INTENT(INOUT) :: res(m,lM%nNo), resE(lM%nEl)
+      REAL(KIND=RKIND), INTENT(IN) :: lD(tDof,tnNo)
+
+      LOGICAL incompFlag
+      INTEGER(KIND=IKIND) a, b, e, g, i, j, k, l, iFn, Ac, nFn, insd,
+     2   eNoN, cPhys, nwg
+      REAL(KIND=RKIND) :: w, nu, Jac0, Jac, Je, lam3, detF, trS, vmises,
+     2   nV0(3), nV(3), aCov0(3,2), aCov(3,2), aCnv0(3,2), aCnv(3,2),
+     3   aa_0(2,2), aa_x(2,2), bb_0(2,2), bb_x(2,2), r0_xx(2,2,3),
+     4   r_xx(2,2,3), Sm(3,2), Dm(3,3,3), Im(3,3), F(3,3), C(3,3),
+     5   Eg(3,3), S(3,3), PK1(3,3), sigma(3,3)
+
+      INTEGER, ALLOCATABLE :: ptr(:)
+      REAL(KIND=RKIND), ALLOCATABLE :: sA(:), sF(:,:), sE(:), resl(:),
+     2   dl(:,:), x0(:,:), xc(:,:), fN(:,:), fNa0(:,:), N(:), Nx(:,:),
+     3   Nxx(:,:), Bb(:,:,:), tmpX(:,:)
+
+      dof  = eq(iEq)%dof
+      i    = eq(iEq)%s
+      j    = i + 1
+      k    = j + 1
+      nFn  = lM%nFn
+      IF (nFn .EQ. 0) nFn = 1
+
+!     Set shell dimension := 2
+      insd = nsd-1
+
+!     Initialize tensor operations
+      CALL TEN_INIT(insd)
+
+!     Set eNoN (number of nodes per element)
+      eNoN = lM%eNoN
+      IF (lM%eType .EQ. eType_TRI3) eNoN = 2*eNoN
+
+!     Allocate arrays
+      ALLOCATE(sA(tnNo), sF(m,tnNo), sE(lM%nEl), resl(m), dl(tDof,eNoN),
+     2   x0(3,eNoN), xc(3,eNoN), fN(3,nFn), fNa0(2,eNoN), ptr(eNoN),
[…]
