+++ Code/Source/svFSI/SHELLS.f
@@ -92,12 +92,13 @@
             yl(:,a)  = Yg(:,Ac)
             dl(:,a)  = Dg(:,Ac)
             bfl(:,a) = Bf(:,Ac)
-            IF (ALLOCATED(lM%fN)) THEN
-               DO iFn=1, nFn
-                  fN(:,iFn) = lM%fN((iFn-1)*nsd+1:iFn*nsd,e)
-               END DO
-            END IF
          END DO
+
+         IF (ALLOCATED(lM%fN)) THEN
+            DO iFn=1, nFn
+               fN(:,iFn) = lM%fN((iFn-1)*nsd+1:iFn*nsd,e)
+            END DO
+         END IF
 
          IF (lM%eType .EQ. eType_TRI3) THEN
 !           Constant strain triangles, no numerical integration
@@ -114,7 +115,7 @@
 !           Gauss integration
             DO g=1, lM%nG
                CALL SHELL3D(lM, g, eNoN, nFn, fN, al, yl, dl, xl,
-     2            bfl, lR, lK)
+     2            bfl, lR, lK, e)
             END DO
 
 !           Assembly
@@ -270,13 +271,12 @@
 !     Compute curvature coefficients for bending strain and its
 !     variation for CST elements
       CALL SHELLBENDCST(lM, e, ptr, x0, xc, bb_0, bb_x, Bb)
-
 !---------------------------------------------------------------------
 !     Compute stress resultants by integrating 2nd Piola Kirchhoff
 !     stress and elasticity tensors through the shell thickness. These
 !     resultants are computed in Voigt notation.
       CALL SHL_STRS_RES(eq(cEq)%dmn(cDmn), nFn, fNa0, aa_0, aa_x, bb_0,
-     2   bb_x, Sm, Dm)
+     2   bb_x, Sm, Dm, e)
 
 !---------------------------------------------------------------------
 !     Contribution to tangent matrices: Dm * Bm, Dm*Bb
@@ -616,7 +616,7 @@
      2         nInI(2,3)*a0(3,p)) + x0(2,i)
             x0(3,j) = 2._RKIND*(nInI(3,1)*a0(1,p) + nInI(3,2)*a0(2,p) +
      2         nInI(3,3)*a0(3,p)) + x0(3,i)
-
+      
 !           Current config
 !           eI = aI/|aI| (current config)
             aIi   = 1._RKIND/SQRT(NORM(a(:,i)))
@@ -704,7 +704,7 @@
       v0(1) = Tm0(1,1)*xi0(3,1) + Tm0(1,2)*xi0(3,2) + Tm0(1,3)*xi0(3,3)
       v0(2) = Tm0(2,1)*xi0(3,1) + Tm0(2,2)*xi0(3,2) + Tm0(2,3)*xi0(3,3)
       v0(3) = Tm0(3,1)*xi0(3,1) + Tm0(3,2)*xi0(3,2) + Tm0(3,3)*xi0(3,3)
-
+ 
 !     Curvature coefficients (ref. config)
       bb_0(1,1) = 2._RKIND*v0(1)
       bb_0(2,2) = 2._RKIND*v0(2)
@@ -959,12 +959,12 @@
 !####################################################################
 !     Construct shell mechanics for higher order elements/NURBS
       SUBROUTINE SHELL3D (lM, g, eNoN, nFn, fN, al, yl, dl, xl, bfl,
-     2   lR, lK)
+     2   lR, lK, e)
       USE COMMOD
       USE MATFUN
       IMPLICIT NONE
[…]
