+++ Code/Source/svFSI/POST.f
@@ -66,7 +66,13 @@
             ALLOCATE(tmpV(1,msh(iM)%nNo), tmpVe(msh(iM)%nEl))
             tmpV  = 0._RKIND
             tmpVe = 0._RKIND
-            CALL TPOST(msh(iM), 1, tmpV, tmpVe, lD, lY, iEq, outGrp)
+            IF (msh(iM)%lShl) THEN
+               CALL SHLTPOST(msh(iM), 1, tmpV, tmpVe, lD, lY, 
+     2          iEq, outGrp)
+            ELSE
+               CALL TPOST(msh(iM), 1, tmpV, tmpVe, lD, lY, iEq,
+     2          outGrp)
+            END IF
             res  = 0._RKIND
             DO a=1, msh(iM)%nNo
                Ac = msh(iM)%gN(a)
@@ -873,29 +879,23 @@
                DO i=1, fs%eNoN
                   resl(:) = resl(:) + N(i)*yl(:,i)
                END DO
-
                IF (eNds(Ac) .EQ. 0) eNds(Ac) = 1
                sF(:,Ac) = sF(:,Ac) + resl(:)*Je
                sA(Ac)   = sA(Ac)   + Je
             END DO
          END DO
-
          CALL COMMU(sF)
          CALL COMMU(sA)
-
          DO a=1, lM%nNo
             Ac = lM%gN(a)
             IF (eNds(Ac).EQ.1 .AND. .NOT.ISZERO(sA(Ac))) THEN
                res(:,a) = res(:,a) + sF(:,Ac)/sA(Ac)
             END IF
          END DO
-
          DEALLOCATE(eNds)
       END IF
-
       DEALLOCATE (sA, sF, sE, xl, dl, yl, tmXl, ya_l, fN, resl, N, Nx)
       CALL DESTROY(fs)
-
       RETURN
       END SUBROUTINE TPOST
 !####################################################################
@@ -1383,4 +1383,584 @@
       STOP
 
       END SUBROUTINE PPBIN2VTK
+!####################################################################
+!     Routine for post processing shell stress tensor
+      SUBROUTINE SHLTPOST(lM, m, res, resE, lD, lY, iEq, outGrp)
+      USE COMMOD
+      USE ALLFUN
+      USE MATFUN
+      IMPLICIT NONE
+      TYPE(mshType), INTENT(INOUT) :: lM
+      INTEGER(KIND=IKIND), INTENT(IN) :: m, iEq, outGrp
+      REAL(KIND=RKIND), INTENT(INOUT) :: res(m,lM%nNo), resE(lM%nEl)
+      REAL(KIND=RKIND), INTENT(IN) :: lD(tDof,tnNo), lY(tDof,tnNo)
+
+      LOGICAL flag
+      INTEGER(KIND=IKIND) a, b, e, g, Ac, i, j, k, l, cPhys, insd, nFn,
+     2   gs, iFn, li, lj, nwg, eNoN 
+      REAL(KIND=RKIND) w, Jac, Jac0, detF, Je, tmX, ya, nu, mu, p, trS,
+     2   vmises, xi(nsd), xi0(nsd), xp(nsd), ed(nsymd), Im(nsd,nsd),
+     3   F(nsd,nsd), C(nsd,nsd), Eg(nsd,nsd), P1(nsd,nsd), S(nsd,nsd),
+     4   sigma(nsd,nsd), aa_0(2,2), aa_x(2,2), aCov0(3,2), aCov(3,2),
+     5   aCnv0(3,2), aCnv(3,2), nV0(3), nV(3), r0_xx(2,2,3), 
[…]
