+++ Code/Source/svFSI/MAIN.f
@@ -153,7 +153,8 @@
             CALL SETBCDIRW(Yg, Dg)
 
 !        Apply contact model and add its contribution to residue
-            IF (iCntct) CALL CONTACTFORCES(Dg)
+            ! IF (iCntct) CALL CONSTRUCT_CONTACTPNLTY(Dg)
+            IF (iCntct) CALL CONSTRUCT_CONTACTPTNL(Dg)
 
 !        Synchronize R across processes. Note: that it is important
 !        to synchronize residue, R before treating immersed bodies as
