### 2022-09-26 94a2ff5
Remove JOSS-related information from main branch.
