+++ paper.bib
@@ -1,225 +1,0 @@
-@article{Xu2020,
-  author={Xu, JQ and Murphy, SL and Kochanek, KD and Arias, E},
-  title={{Mortality in the United States, 2018}},
-  journal={NCHS Data Brief No. 355},
-  year={2020},
-  publisher={National Center for Health Statistics},
-  url = {https://www.cdc.gov/nchs/products/databriefs/db355.htm}
-}
-
-@article{Mittal2015,
-  author = {Mittal, Rajat and Seo, Jung Hee and Vedula, Vijay and Choi, Young J. and Liu, Hang and Huang, H. Howie and Jain, Saurabh and Younes, Laurent and Abraham, Theodore and George, Richard T.},
-  doi = {10.1016/j.jcp.2015.11.022},
-  issn = {00219991},
-  journal = {Journal of Computational Physics},
-  pages = {1065--1082},
-  publisher = {Elsevier Inc.},
-  title = {{Computational modeling of cardiac hemodynamics: Current status and future outlook}},
-  url = {http://linkinghub.elsevier.com/retrieve/pii/S0021999115007627},
-  volume = {305},
-  year = {2015}
-}
-
-@article{Trayanova2011,
-author = {Trayanova, Natalia A.},
-doi = {10.1161/CIRCRESAHA.110.223610},
-issn = {00097330},
-journal = {Circulation Research},
-keywords = {Cardiac disease,Electromechanical modeling,Electrophysiological modeling,Simulation,Whole-heart model},
-number = {1},
-pages = {113--128},
-pmid = {21212393},
-title = {{Whole-heart modeling : Applications to cardiac electrophysiology and electromechanics}},
-volume = {108},
-year = {2011}
-}
-
-
-@article{Updegrove2017,
-  author = {Updegrove, Adam and Wilson, Nathan M. and Merkow, Jameson and Lan, Hongzhi and Marsden, Alison L. and Shadden, Shawn C.},
-  doi = {10.1007/s10439-016-1762-8},
-  issn = {15739686},
-  journal = {Annals of Biomedical Engineering},
-  keywords = {Hemodynamics,Image-based CFD,Open-source,Patient-specific modeling},
-  number = {3},
-  pages = {525--541},
-  pmid = {27933407},
-  title = {{SimVascular: An Open Source Pipeline for Cardiovascular Simulation}},
-  volume = {45},
-  year = {2017}
-}
-
-@article{Maas2012,
-  title={FEBio: finite elements for biomechanics},
-  author={Maas, Steve A and Ellis, Benjamin J and Ateshian, Gerard A and Weiss, Jeffrey A},
-  journal={Journal of biomechanical engineering},
-  volume={134},
-  number={1},
-  year={2012},
-  doi={10.1115/1.4005694},
-  publisher={American Society of Mechanical Engineers Digital Collection}
-}
-
-@article{Plank2021,
-title = {The openCARP simulation environment for cardiac electrophysiology},
-journal = {Computer Methods and Programs in Biomedicine},
-volume = {208},
-pages = {106223},
-year = {2021},
-issn = {0169-2607},
-doi = {https://doi.org/10.1016/j.cmpb.2021.106223},
-url = {https://www.sciencedirect.com/science/article/pii/S0169260721002972},
-author = {Gernot Plank and Axel Loewe and Aurel Neic and Christoph Augustin and Yung-Lin Huang and Matthias A.F. Gsell and Elias Karabelas and Mark Nothstein and Anton J. Prassl and Jorge Sánchez and Gunnar Seemann and Edward J. Vigmond},
-}
-
-@article{Bertagna2017,
-  title={The LifeV library: engineering mathematics beyond the proof of concept},
[…]
