### 2022-09-26 #4118 @CZHU20
@Kevin-Mattheus-Moerman
Dear Kevin, I have made all the changes. We have created a new release (2022.09.26) based on the reviewed version of svFSI. The DOI of the ZENODO archive is [10.5281/zenodo.6673060](https://zenodo.org/record/7113485#.YzGv0C0RqZw). All records on ZENODO has been made consistent with the JOSS submission. Please let me know if there are other issues. Thank you very much.
