### 2022-10-02 8a001a5
updating shell formulas
