+++ Code/Source/svFSI/MATMODELS.f
@@ -1323,7 +1323,9 @@ c     2      (EXP(stM%khs*Ess) + EXP(-stM%khs*Ess) + 2.0_RKIND)
       REAL(KIND=RKIND) :: Jg2, J2, J23, f13, f23, trC3, C33, kap, mu,
      2   pJ, plJ, gi_x(2,2), gi_0(3,3), Ci(3,3), S(3,3), CC(3,3,3,3),
      3   fl(2,2,nfd), C1, C2, J43, Gi4AS(3,3,3,3), I2, I2ij(3,3),
-     4   I2ijkl(3,3,3,3), Cikl(3,3,3,3)
+     4   I2ijkl(3,3,3,3), Cikl(3,3,3,3), Inv4, Inv6, Inv8, Eff, Ess,
+     5   Efs, c4f, c4s, dc4f, dc4s, EI1, d1, g1, g2, Hfs(3,3), 
+     6   SN(3,3), CCN(3,3,3,3), flM(3,3)
       TYPE(stModelType) :: stM
 
       Sml  = 0._RKIND
@@ -1410,24 +1412,122 @@ c     2      (EXP(stM%khs*Ess) + EXP(-stM%khs*Ess) + 2.0_RKIND)
      3                - TEN_DYADPROD(Ci, I2ij, 3) + I2ijkl)
                      
          CASE (stIso_HO_ma)
-!           2nd Piola Kirchhoff stress
-            S  = mu*J23*(gi_0 - trC3*Ci) + pJ*Ci
+            IF (nfd .NE. 2) err = "Min fiber directions not defined"//  
+     2          "for Holzapfel material model (2)"
 
-!           Elasticity tensor
-            CC = (mu*J23*f23*trC3 + plJ)*TEN_DYADPROD(Ci, Ci, 3)
-     2         + (mu*J23*trC3 - pJ)*2._RKIND*TEN_SYMMPROD(Ci, Ci, 3)
-     3         - f23*mu*J23*(TEN_DYADPROD(gi_0, Ci, 3) +
-     4                       TEN_DYADPROD(Ci, gi_0, 3))
+            DO iFn=1, nfd
+                fl(1,1,iFn) = fNa0(1,iFn)*fNa0(1,iFn)
+                fl(1,2,iFn) = fNa0(1,iFn)*fNa0(2,iFn)
+                fl(2,1,iFn) = fNa0(2,iFn)*fNa0(1,iFn)
+                fl(2,2,iFn) = fNa0(2,iFn)*fNa0(2,iFn)
+            END DO
+
+    !        Compute fiber-based invariants
+            Inv4 = gg_x(1,1)*fl(1,1,1) + gg_x(1,2)*fl(1,2,1)
+     2              + gg_x(2,1)*fl(2,1,1) + gg_x(2,2)*fl(2,2,1)
+            Inv6 = gg_x(1,1)*fl(1,1,2) + gg_x(1,2)*fl(1,2,2)
+     2            + gg_x(2,1)*fl(2,1,2) + gg_x(2,2)*fl(2,2,2)
+            Inv8 = gg_x(1,1)*fNa0(1,1)*fNa0(1,2)
+     2              + gg_x(1,2)*fNa0(1,1)*fNa0(2,2)
+     3              + gg_x(2,1)*fNa0(2,1)*fNa0(1,2)
+     4              + gg_x(2,2)*fNa0(2,1)*fNa0(2,2)
+
+            Eff  = Inv4 - 1._RKIND
+            Ess  = Inv6 - 1._RKIND
+            Efs  = Inv8
+
+    !        Smoothed heaviside function
+            c4f  = 1._RKIND / (1._RKIND + EXP(-stM%khs*Eff))
+            c4s  = 1._RKIND / (1._RKIND + EXP(-stM%khs*Ess))
+
[…]
