+++ Code/Source/svFSI/CONSTS.f
@@ -159,7 +159,8 @@
      5   outGrp_absV = 512, outGrp_fN = 513, outGrp_fA = 514,
      6   outGrp_stress = 515, outGrp_cauchy = 516, outGrp_mises = 517,
      7   outGrp_J = 518, outGrp_F = 519, outGrp_strain = 520,
-     8   outGrp_divV = 521, outGrp_Visc = 522, outGrp_fS = 523
+     8   outGrp_divV = 521, outGrp_Visc = 522, outGrp_fS = 523, 
+     9   outGrp_C = 524, outGrp_I1 = 525
 !--------------------------------------------------------------------
       INTEGER(KIND=IKIND), PARAMETER :: out_velocity = 599,
      2   out_pressure = 598, out_temperature = 597, out_voltage = 596,
@@ -170,7 +171,7 @@
      7   out_fibAlign = 583, out_stress = 582, out_cauchy = 581,
      8   out_mises = 580, out_jacobian = 579, out_defGrad = 578,
      9   out_strain = 577, out_divergence = 576, out_viscosity = 575,
-     1   out_fibStrn = 574
+     1   out_fibStrn = 574, out_CGstrain = 573, out_CGInv1 = 572
 !--------------------------------------------------------------------
 !     Mesher choice for remeshing for moving wall problems
       INTEGER(KIND=IKIND), PARAMETER :: RMSH_TETGEN = 1,
