+++ Code/Source/svFSI/POST.f
@@ -514,7 +514,7 @@
       REAL(KIND=RKIND) w, Jac, detF, Je, tmX, ya, Ja, elM, nu, lambda,
      2   mu, p, trS, vmises, xi(nsd), xi0(nsd), xp(nsd), ed(nsymd),
      3   Im(nsd,nsd), F(nsd,nsd), C(nsd,nsd), Eg(nsd,nsd), P1(nsd,nsd),
-     4   S(nsd,nsd), sigma(nsd,nsd), Dm(nsymd,nsymd)
+     4   S(nsd,nsd), sigma(nsd,nsd), Dm(nsymd,nsymd), I1
       TYPE(fsType) :: fs
 
       INTEGER, ALLOCATABLE :: eNds(:)
@@ -689,27 +689,46 @@
                   resl(4) = F(2,2)
                END IF
 
-            CASE (outGrp_strain)
-!           Green-Lagrange strain tensor
-               IF (cPhys .EQ. phys_lElas) THEN
-                  resl(:) = ed(:)
-               ELSE
-                  C  = MATMUL(TRANSPOSE(F), F)
-                  Eg = 0.5_RKIND * (C - Im)
-                  ! resl is used to remap Eg
-                  IF (nsd .EQ. 3) THEN
-                     resl(1) = Eg(1,1)
-                     resl(2) = Eg(2,2)
-                     resl(3) = Eg(3,3)
-                     resl(4) = Eg(1,2)
-                     resl(5) = Eg(2,3)
-                     resl(6) = Eg(3,1)
+            CASE (outGrp_strain, outGrp_C, outGrp_I1)
+               IF (outGrp .EQ. outGrp_strain) THEN
+!              Green-Lagrange strain tensor
+                  IF (cPhys .EQ. phys_lElas) THEN
+                     resl(:) = ed(:)
                   ELSE
-                     resl(1) = Eg(1,1)
-                     resl(2) = Eg(2,2)
-                     resl(3) = Eg(1,2)
+                     C  = MATMUL(TRANSPOSE(F), F)
+                     Eg = 0.5_RKIND * (C - Im)
+                     ! resl is used to remap Eg
+                     IF (nsd .EQ. 3) THEN
+                        resl(1) = Eg(1,1)
+                        resl(2) = Eg(2,2)
+                        resl(3) = Eg(3,3)
+                        resl(4) = Eg(1,2)
+                        resl(5) = Eg(2,3)
+                        resl(6) = Eg(3,1)
+                     ELSE
+                        resl(1) = Eg(1,1)
+                        resl(2) = Eg(2,2)
+                        resl(3) = Eg(1,2)
+                     END IF
                   END IF
-               END IF
+               ELSE IF (outGrp .EQ. outGrp_C) THEN
+                  C  = MATMUL(TRANSPOSE(F), F)
+                  IF (nsd .EQ. 3) THEN
+                     resl(1) = C(1,1)
+                     resl(2) = C(2,2)
+                     resl(3) = C(3,3)
+                     resl(4) = C(1,2)
+                     resl(5) = C(2,3)
+                     resl(6) = C(3,1)
+                  ELSE
+                     resl(1) = C(1,1)
+                     resl(2) = C(2,2)
+                     resl(3) = C(1,2)
[…]
