### 2022-10-02 02ec97d
Cleaned up code while merging with the main branch
