+++ Code/Source/svFSI/MATMODELS.f
@@ -1306,7 +1306,7 @@ c     2      (EXP(stM%khs*Ess) + EXP(-stM%khs*Ess) + 2.0_RKIND)
 !--------------------------------------------------------------------
 !     Compute 2nd Piola-Kirchhoff stress and material stiffness tensors
 !     for compressible shell elements
-      SUBROUTINE GETPK2CC_SHLc(lDmn, nfd, fNa0, gg_0, gg_x, g33, Sml, 
+      SUBROUTINE GETPK2CC_SHLc(lDmn, nfd, fNa0, gg_0, gg_x, g33, Sml,
      2  Dml)
       USE MATFUN
       USE COMMOD
@@ -1324,8 +1324,8 @@ c     2      (EXP(stM%khs*Ess) + EXP(-stM%khs*Ess) + 2.0_RKIND)
      2   pJ, plJ, gi_x(2,2), gi_0(3,3), Ci(3,3), S(3,3), CC(3,3,3,3),
      3   fl(2,2,nfd), C1, C2, J43, Gi4AS(3,3,3,3), I2, I2ij(3,3),
      4   I2ijkl(3,3,3,3), Cikl(3,3,3,3), Inv4, Inv6, Inv8, Eff, Ess,
-     5   Efs, c4f, c4s, dc4f, dc4s, EI1, d1, g1, g2, Hfs(3,3), 
-     6   SN(3,3), CCN(3,3,3,3), flM(3,3)
+     5   Efs, c4f, c4s, dc4f, dc4s, d1, g1, g2, Hfs(3,3), SN(3,3),
+     6   CCN(3,3,3,3), flM(3,3)
       TYPE(stModelType) :: stM
 
       Sml  = 0._RKIND
@@ -1388,31 +1388,34 @@ c     2      (EXP(stM%khs*Ess) + EXP(-stM%khs*Ess) + 2.0_RKIND)
 
          CASE (stIso_MR)
 !           2nd Piola Kirchhoff stress
-            C1 = stM%C10
-            C2 = stM%C01
+            C1  = stM%C10
+            C2  = stM%C01
             J43 = J2**(-f23)
-            I2ijkl = TEN_DYADPROD(gi_0, gi_0, 3) - 
-     2                              TEN_SYMMPROD(gi_0, gi_0, 3)          
-            I2ij = TEN_MDDOT(I2ijkl, Ci, 3)
-            Gi4AS = TEN_ASYMPROD12(gi_0, gi_0, 3)
-            I2  = MAT_DDOT(Ci, TEN_MDDOT(Gi4AS, Ci, 3), 3)
+
+            I2ijkl = TEN_DYADPROD(gi_0, gi_0, 3)
+     2             - TEN_SYMMPROD(gi_0, gi_0, 3)
+            I2ij   = TEN_MDDOT(I2ijkl, Ci, 3)
+
+            Gi4AS  = TEN_ASYMPROD12(gi_0, gi_0, 3)
+            I2     = MAT_DDOT(Ci, TEN_MDDOT(Gi4AS, Ci, 3), 3)
             ! I2  = MAT_DDOT(Ci, I2ij, 3)
-            Cikl = - TEN_SYMMPROD(Ci, Ci, 3)
+            Cikl   = - TEN_SYMMPROD(Ci, Ci, 3)
 
             S  = C1*J23*(gi_0 - trC3*Ci) + pJ*Ci
-            S  = S + C2*J43*(-f23*I2*Ci + I2ij)
+     2         + C2*J43*(I2ij - f23*I2*Ci)
 
 !           Elasticity tensor
[…]
