### 2022-10-03 38f0ffe
Fixing mem bug with ustruct with cap.
