### 2022-10-10 18e723a
Added DOI entry in JOSS paper bib file
Added `doi = {10.48550/arXiv.1710.06596}` for Bertagna2017 entry
