+++ paper.bib
@@ -77,6 +77,7 @@ author = {Gernot Plank and Axel Loewe and Aurel Neic and Christoph Augustin and 
   author={Bertagna, Luca and Deparis, Simone and Formaggia, Luca and Forti, Davide and Veneziani, Alessandro},
   journal={arXiv preprint},
   year={2017},
+  doi = {10.48550/arXiv.1710.06596},  
   url={https://arxiv.org/abs/1710.06596}
 }
 
